==> 03.1_Physiology.Rmd <==
---
title: "Coral Physiology Following Reciprocal Transplant"
author: "Serena Hackerott"
date: "1/1/2024"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  html_document:
    toc: yes
    df_print: paged
---

# Setup

```{r Setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


### Load Packages
```{r}
##Install Packages if Needed
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("cowplot")) install.packages("cowplot")
if (!require("Rmisc")) install.packages("Rmisc")
if (!require("lme4")) install.packages("lme4")
if (!require("lmerTest")) install.packages("lmerTest")
if (!require("vegan")) install.packages("vegan")
if (!require("corrplot")) install.packages("corrplot")
if (!require("DHARMa")) install.packages("DHARMa")
if (!require("effectsize")) install.packages("effectsize")
if (!require("emmeans")) install.packages("emmeans")
if (!require("Hmisc")) install.packages("Hmisc")
if (!require("dplyr")) install.packages("dplyr")
if (!require("tidyr")) install.packages("tidyr")
if (!require("ggpubr")) install.packages("ggpubr")

##Load Packages
library(ggplot2) #Required for ggplots
library(cowplot) #Required for plotting panel figures
library(Rmisc) #Required for SummarySE function
library(lme4) #Required for mixed effects modeling
library(lmerTest) #Required for p values with lme4 model summaries
library(vegan) #Required for multivariate analysis PERM
library(corrplot) #Required for correlation plot
library(DHARMa) #Required to check residuals of mixed effects models
library(effectsize) #Required for eta_squared effect sizes
library(emmeans) #Required for pairwise comparisons 
library(Hmisc) #Required for correlations
library(dplyr) #Required for dataframe organization
library(tidyr) #Required for dataframe organization
library(ggpubr) #Required for stat brackets

```


### Graphing Parameters
```{r}
#Note: Run "Graphing Parameters" section from 01_ExperimentalSetup.Rmd file
```


# Sample Data and Metadata

### Load and Organize Data
```{r}
##Load Data
#Note: Physiological metrics calculated in 02_PhysiologyMetrics.R file
Coral<-read.csv("Outputs/CoralData.csv", header=TRUE)

##Set factor variables
Coral$TimeP<-factor(Coral$TimeP, levels=c("TP1", "TP2", "TP3", "TP4"))
Coral$Site<-factor(Coral$Site, levels=c("KL", "SS"))
Coral$Genotype<-factor(Coral$Genotype, levels=c("AC8", "AC10", "AC12"))
Coral$Orig<-factor(Coral$Orig, levels=c("N", "T"))
Coral$Origin<-factor(Coral$Origin, levels=c("Native", "Transplant"))
Coral$Site.Orig<-factor(Coral$Site.Orig, levels=c("KL.N", "KL.T", "SS.N", "SS.T"))


```


# Multivariate Analysis

### adonis OmegaSq Function
```{r}
#' Calculate (partial) Omega-squared (effect-size calculation) for PERMANOVA and add it to the input object
#'
#' @param adonisOutput An adonis object
#' @param partial Should partial omega-squared be calculated (sample size adjusted). Default TRUE
#' @return Original adonis object with the (partial) Omega-squared values added
#' @import vegan
#' @export
adonis_OmegaSq <- function(adonisOutput, partial = TRUE){
    if(!(is(adonisOutput, "adonis") || is(adonisOutput, "anova.cca")))
        stop("Input should be an adonis object")
    if (is(adonisOutput, "anova.cca")) {
        aov_tab <- adonisOutput
        aov_tab$MeanSqs <- aov_tab$SumOfSqs / aov_tab$Df
        aov_tab$MeanSqs[length(aov_tab$Df)] <- NA
    } else {
        aov_tab <- adonisOutput$aov.tab
    }
    heading <- attr(aov_tab, "heading")
    MS_res <- aov_tab[pmatch("Residual", rownames(aov_tab)), "MeanSqs"]
    SS_tot <- aov_tab[rownames(aov_tab) == "Total", "SumsOfSqs"]
    N <- aov_tab[rownames(aov_tab) == "Total", "Df"] + 1
    if(partial){
        omega <- apply(aov_tab, 1, function(x) (x["Df"]*(x["MeanSqs"]-MS_res))/(x["Df"]*x["MeanSqs"]+(N-x["Df"])*MS_res))
        aov_tab$parOmegaSq <- c(omega[1:(length(omega)-2)], NA, NA)
    } else {
        omega <- apply(aov_tab, 1, function(x) (x["SumsOfSqs"]-x["Df"]*MS_res)/(SS_tot+MS_res))
        aov_tab$OmegaSq <- c(omega[1:(length(omega)-2)], NA, NA)
    }
    if (is(adonisOutput, "adonis"))
        cn_order <- c("Df", "SumsOfSqs", "MeanSqs", "F.Model", "R2",
                      if (partial) "parOmegaSq" else "OmegaSq", "Pr(>F)")
    else
        cn_order <- c("Df", "SumOfSqs", "F", if (partial) "parOmegaSq" else "OmegaSq",
                      "Pr(>F)")
    aov_tab <- aov_tab[, cn_order]
    attr(aov_tab, "names") <- cn_order
    attr(aov_tab, "heading") <- heading
    if (is(adonisOutput, "adonis"))
        adonisOutput$aov.tab <- aov_tab
    else
        adonisOutput <- aov_tab
    return(adonisOutput)
}
```


### Check Correlation
```{r}
##Remove NA's
names(Coral)
Coral.rm<-na.omit(Coral)

##Correlation 
Phys.cor<-rcorr(as.matrix(Coral.rm[,-c(1:10)]), type="pearson")
Phys.cor

diag(Phys.cor$P)<-0

corrplot(Phys.cor$r, type="upper", order="hclust", 
         p.mat = Phys.cor$P, sig.level = 0.01, insig = "blank")

```

### All Timepoints
```{r}
##Standardize Variables for similar scales
Coral_St <- Coral.rm %>% mutate_at(c("TP_ug.cm2", "AFDW_mg.cm2", "Chl_ug.cm2"), ~(scale(.) %>% as.vector))

```


#### PERMANOVA
Checking for an overall effect of Origin or Site or Time, along with all interactions. Overall effect of Time alone (season) is not of main interest, but is predicted to interact with the effect of Origin as the time post-transplant increases.  
```{r}
##PERMANOVA
Coral.perm<-adonis2(vegdist(Coral_St[,c(11:13)], "euclidean")~ Coral_St$Origin*Coral_St$Site*Coral_St$TimeP, data=Coral_St, strata=Coral_St$Genotype, method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.perm)

##Check dispersion by Origin
Orig_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Origin))
Orig_disp

##Check dispersion by Site
Site_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Site))
Site_disp

##Check dispersion by Time
Time_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$TimeP))
Time_disp

##Check dispersion by Origin:Site
Orig.Site_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Origin:Coral_St$Site))
Orig.Site_disp

##Check dispersion by Origin:Time
Orig.Time_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Origin:Coral_St$TimeP))
Orig.Time_disp

##Check dispersion by Site:Time
Site.Time_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Site:Coral_St$TimeP))
Site.Time_disp

##Check dispersion by Origin:Site:Time
Orig.Site.Time_disp<-anova(betadisper(vegdist(Coral_St[,c(11:13)], "euclidean"), Coral_St$Origin:Coral_St$Site:Coral_St$TimeP))
Orig.Site.Time_disp

##Save results
PERM.res<-data.frame(adonis_OmegaSq(Coral.perm)[1:7,])
PERM.res$Predictor<-c("Origin", "Site", "Time", "Origin x Site", "Origin x Time", "Site x Time", "Site x Time x Origin")
PERM.res$p_DISP<-c(Orig_disp$`Pr(>F)`[1], Site_disp$`Pr(>F)`[1], Time_disp$`Pr(>F)`[1],
                   Orig.Site_disp$`Pr(>F)`[1], Orig.Time_disp$`Pr(>F)`[1], 
                   Site.Time_disp$`Pr(>F)`[1], Orig.Site.Time_disp$`Pr(>F)`[1])
PERM.res$Timepoint<-rep("All", nrow(PERM.res))

```

Due to significant interactions with Time, each time point will be analyzed separately to assess the effect of origin vs current site. 


### TP1
```{r}
##Subset Timepoint 1 
Coral_TP1<-subset(Coral.rm, TimeP=="TP1")

##Standardize Variables for similar scales
Coral_TP1 <- Coral_TP1 %>% mutate_at(c("TP_ug.cm2", "AFDW_mg.cm2", "Chl_ug.cm2"), ~(scale(.) %>% as.vector))

```


#### PERMANOVA
```{r}
##PERMANOVA
Coral.TP1.perm<-adonis2(vegdist(Coral_TP1[,c(11:13)], "euclidean")~ Coral_TP1$Origin * Coral_TP1$Site, data=Coral_TP1, strata=Coral_TP1$Genotype, method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP1.perm)

##Check dispersion by Origin
TP1.disp_orig<-anova(betadisper(vegdist(Coral_TP1[,c(11:13)], "euclidean"), Coral_TP1$Origin))
TP1.disp_orig

##Check dispersion by Site
TP1.disp_site<-anova(betadisper(vegdist(Coral_TP1[,c(11:13)], "euclidean"), Coral_TP1$Site))
TP1.disp_site

##Check dispersion by Origin:Site
TP1.disp_os<-anova(betadisper(vegdist(Coral_TP1[,c(11:13)], "euclidean"), Coral_TP1$Origin:Coral_TP1$Site))
TP1.disp_os

##Save results
PERM_T1.res<-data.frame(adonis_OmegaSq(Coral.TP1.perm)[1:3,])
PERM_T1.res$Predictor<-c("Origin", "Site", "Origin x Site")
PERM_T1.res$p_DISP<-c(TP1.disp_orig$`Pr(>F)`[1], TP1.disp_site$`Pr(>F)`[1], TP1.disp_os$`Pr(>F)`[1])
```
Physiology differs significantly by Origin and the effect of Origin differs between Sites.

#### Effect Size
Calculate Effect Size of Origin for each Site
```{r}
##KL
##PERMANOVA
Coral.TP1.perm_KL<-adonis2(vegdist(Coral_TP1[which(Coral_TP1$Site=="KL"),c(11:13)], "euclidean")~ Coral_TP1$Origin[which(Coral_TP1$Site=="KL")], data=Coral_TP1[which(Coral_TP1$Site=="KL"),], strata=Coral_TP1$Genotype[which(Coral_TP1$Site=="KL")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP1.perm_KL)

##Check dispersion by Origin
TP1.KL.disp_orig<-anova(betadisper(vegdist(Coral_TP1[which(Coral_TP1$Site=="KL"),c(11:13)], "euclidean"), Coral_TP1$Origin[which(Coral_TP1$Site=="KL")]))
TP1.KL.disp_orig

##SS
##PERMANOVA
Coral.TP1.perm_SS<-adonis2(vegdist(Coral_TP1[which(Coral_TP1$Site=="SS"),c(11:13)], "euclidean")~ Coral_TP1$Origin[which(Coral_TP1$Site=="SS")], data=Coral_TP1[which(Coral_TP1$Site=="SS"),], strata=Coral_TP1$Genotype[which(Coral_TP1$Site=="SS")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP1.perm_SS)

##Check dispersion by Origin
TP1.SS.disp_orig<-anova(betadisper(vegdist(Coral_TP1[which(Coral_TP1$Site=="SS"),c(11:13)], "euclidean"), Coral_TP1$Origin[which(Coral_TP1$Site=="SS")]))
TP1.SS.disp_orig

##Save results
PERM_T1.site.res<-data.frame(adonis_OmegaSq(Coral.TP1.perm_KL)[1,])
PERM_T1.site.res<-rbind(PERM_T1.site.res, data.frame(adonis_OmegaSq(Coral.TP1.perm_SS)[1,]))
PERM_T1.site.res$Predictor<-c("KL Origin", "SS Origin")
PERM_T1.site.res$p_DISP<-c(TP1.KL.disp_orig$`Pr(>F)`[1], TP1.SS.disp_orig$`Pr(>F)`[1])

##Combine results 
PERM_T1.res<-rbind(PERM_T1.res, PERM_T1.site.res)
PERM_T1.res$Timepoint<-rep("TP1", nrow(PERM_T1.res))

```


### TP2
```{r}
##Subset Timepoint 2 
Coral_TP2<-subset(Coral.rm, TimeP=="TP2")

##Standardize Variables for similar scales
Coral_TP2 <- Coral_TP2 %>% mutate_at(c("TP_ug.cm2", "AFDW_mg.cm2", "Chl_ug.cm2"), ~(scale(.) %>% as.vector))

```

#### PERMANOVA
```{r}
##PERMANOVA
Coral.TP2.perm<-adonis2(vegdist(Coral_TP2[,c(11:13)], "euclidean")~ Coral_TP2$Origin * Coral_TP2$Site, data=Coral_TP2, strata=Coral_TP2$Genotype, method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP2.perm)

##Check dispersion by Origin
TP2.disp_orig<-anova(betadisper(vegdist(Coral_TP2[,c(11:13)], "euclidean"), Coral_TP2$Origin))
TP2.disp_orig

##Check dispersion by Site
TP2.disp_site<-anova(betadisper(vegdist(Coral_TP2[,c(11:13)], "euclidean"), Coral_TP2$Site))
TP2.disp_site

##Check dispersion by Origin:Site
TP2.disp_os<-anova(betadisper(vegdist(Coral_TP2[,c(11:13)], "euclidean"), Coral_TP2$Origin:Coral_TP2$Site))
TP2.disp_os

##Save results
PERM_T2.res<-data.frame(adonis_OmegaSq(Coral.TP2.perm)[1:3,])
PERM_T2.res$Predictor<-c("Origin", "Site", "Origin x Site")
PERM_T2.res$p_DISP<-c(TP2.disp_orig$`Pr(>F)`[1], TP2.disp_site$`Pr(>F)`[1], TP2.disp_os$`Pr(>F)`[1])

```


#### Effect Size
Calculate Effect Size of Origin for each Site
```{r}
##KL
##PERMANOVA
Coral.TP2.perm_KL<-adonis2(vegdist(Coral_TP2[which(Coral_TP2$Site=="KL"),c(11:13)], "euclidean")~ Coral_TP2$Origin[which(Coral_TP2$Site=="KL")], data=Coral_TP2[which(Coral_TP2$Site=="KL"),], strata=Coral_TP2$Genotype[which(Coral_TP2$Site=="KL")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP2.perm_KL)

##Check dispersion by Origin
TP2.KL.disp_orig<-anova(betadisper(vegdist(Coral_TP2[which(Coral_TP2$Site=="KL"),c(11:13)], "euclidean"), Coral_TP2$Origin[which(Coral_TP2$Site=="KL")]))
TP2.KL.disp_orig

##SS
##PERMANOVA
Coral.TP2.perm_SS<-adonis2(vegdist(Coral_TP2[which(Coral_TP2$Site=="SS"),c(11:13)], "euclidean")~ Coral_TP2$Origin[which(Coral_TP2$Site=="SS")], data=Coral_TP2[which(Coral_TP2$Site=="SS"),], strata=Coral_TP2$Genotype[which(Coral_TP2$Site=="SS")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP2.perm_SS)

##Check dispersion by Origin
TP2.SS.disp_orig<-anova(betadisper(vegdist(Coral_TP2[which(Coral_TP2$Site=="SS"),c(11:13)], "euclidean"), Coral_TP2$Origin[which(Coral_TP2$Site=="SS")]))
TP2.SS.disp_orig

##Save results
PERM_T2.site.res<-data.frame(adonis_OmegaSq(Coral.TP2.perm_KL)[1,])
PERM_T2.site.res<-rbind(PERM_T2.site.res, data.frame(adonis_OmegaSq(Coral.TP2.perm_SS)[1,]))
PERM_T2.site.res$Predictor<-c("KL Origin", "SS Origin")
PERM_T2.site.res$p_DISP<-c(TP2.KL.disp_orig$`Pr(>F)`[1], TP2.SS.disp_orig$`Pr(>F)`[1])

##Combine results 
PERM_T2.res<-rbind(PERM_T2.res, PERM_T2.site.res)
PERM_T2.res$Timepoint<-rep("TP2", nrow(PERM_T2.res))

```


### TP3
```{r}
##Subset Timepoint 3 
Coral_TP3<-subset(Coral.rm, TimeP=="TP3")

##Standardize Variables for similar scales
Coral_TP3 <- Coral_TP3 %>% mutate_at(c("TP_ug.cm2", "AFDW_mg.cm2", "Chl_ug.cm2"), ~(scale(.) %>% as.vector))

```


#### PERMANOVA
```{r}
##PERMANOVA
Coral.TP3.perm<-adonis2(vegdist(Coral_TP3[,c(11:13)], "euclidean")~ Coral_TP3$Origin * Coral_TP3$Site, data=Coral_TP3, strata=Coral_TP3$Genotype, method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP3.perm)

##Check dispersion by Origin
TP3.disp_orig<-anova(betadisper(vegdist(Coral_TP3[,c(11:13)], "euclidean"), Coral_TP3$Origin))
TP3.disp_orig

##Check dispersion by Site
TP3.disp_site<-anova(betadisper(vegdist(Coral_TP3[,c(11:13)], "euclidean"), Coral_TP3$Site))
TP3.disp_site

##Check dispersion by Origin:Site
TP3.disp_os<-anova(betadisper(vegdist(Coral_TP3[,c(11:13)], "euclidean"), Coral_TP3$Origin:Coral_TP3$Site))
TP3.disp_os

##Save results
PERM_T3.res<-data.frame(adonis_OmegaSq(Coral.TP3.perm)[1:3,])
PERM_T3.res$Predictor<-c("Origin", "Site", "Origin x Site")
PERM_T3.res$p_DISP<-c(TP3.disp_orig$`Pr(>F)`[1], TP3.disp_site$`Pr(>F)`[1], TP3.disp_os$`Pr(>F)`[1])

```


#### Effect Size
Calculate Effect Size of Origin for each Site
```{r}
##KL
##PERMANOVA
Coral.TP3.perm_KL<-adonis2(vegdist(Coral_TP3[which(Coral_TP3$Site=="KL"),c(11:13)], "euclidean")~ Coral_TP3$Origin[which(Coral_TP3$Site=="KL")], data=Coral_TP3[which(Coral_TP3$Site=="KL"),], strata=Coral_TP3$Genotype[which(Coral_TP3$Site=="KL")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP3.perm_KL)

##Check dispersion by Origin
TP3.KL.disp_orig<-anova(betadisper(vegdist(Coral_TP3[which(Coral_TP3$Site=="KL"),c(11:13)], "euclidean"), Coral_TP3$Origin[which(Coral_TP3$Site=="KL")]))
TP3.KL.disp_orig

##SS
##PERMANOVA
Coral.TP3.perm_SS<-adonis2(vegdist(Coral_TP3[which(Coral_TP3$Site=="SS"),c(11:13)], "euclidean")~ Coral_TP3$Origin[which(Coral_TP3$Site=="SS")], data=Coral_TP3[which(Coral_TP3$Site=="SS"),], strata=Coral_TP3$Genotype[which(Coral_TP3$Site=="SS")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP3.perm_SS)

##Check dispersion by Origin
TP3.SS.disp_orig<-anova(betadisper(vegdist(Coral_TP3[which(Coral_TP3$Site=="SS"),c(11:13)], "euclidean"), Coral_TP3$Origin[which(Coral_TP3$Site=="SS")]))
TP3.SS.disp_orig

##Save results
PERM_T3.site.res<-data.frame(adonis_OmegaSq(Coral.TP3.perm_KL)[1,])
PERM_T3.site.res<-rbind(PERM_T3.site.res, data.frame(adonis_OmegaSq(Coral.TP3.perm_SS)[1,]))
PERM_T3.site.res$Predictor<-c("KL Origin", "SS Origin")
PERM_T3.site.res$p_DISP<-c(TP3.KL.disp_orig$`Pr(>F)`[1], TP3.SS.disp_orig$`Pr(>F)`[1])

##Combine results 
PERM_T3.res<-rbind(PERM_T3.res, PERM_T3.site.res)
PERM_T3.res$Timepoint<-rep("TP3", nrow(PERM_T3.res))


```


### TP4
```{r}
##Subset Timepoint 4 
Coral_TP4<-subset(Coral.rm, TimeP=="TP4")

##Standardize Variables for similar scales
Coral_TP4 <- Coral_TP2 %>% mutate_at(c("TP_ug.cm2", "AFDW_mg.cm2", "Chl_ug.cm2"), ~(scale(.) %>% as.vector))

```


#### PERMANOVA
```{r}
##PERMANOVA
Coral.TP4.perm<-adonis2(vegdist(Coral_TP4[,c(11:13)], "euclidean")~ Coral_TP4$Origin * Coral_TP4$Site, data=Coral_TP4, strata=Coral_TP4$Genotype, method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP4.perm)

##Check dispersion by Origin
TP4.disp_orig<-anova(betadisper(vegdist(Coral_TP4[,c(11:13)], "euclidean"), Coral_TP4$Origin))
TP4.disp_orig

##Check dispersion by Site
TP4.disp_site<-anova(betadisper(vegdist(Coral_TP4[,c(11:13)], "euclidean"), Coral_TP4$Site))
TP4.disp_site

##Check dispersion by Origin:Site
TP4.disp_os<-anova(betadisper(vegdist(Coral_TP4[,c(11:13)], "euclidean"), Coral_TP4$Origin:Coral_TP4$Site))
TP4.disp_os

##Save results
PERM_T4.res<-data.frame(adonis_OmegaSq(Coral.TP4.perm)[1:3,])
PERM_T4.res$Predictor<-c("Origin", "Site", "Origin x Site")
PERM_T4.res$p_DISP<-c(TP4.disp_orig$`Pr(>F)`[1], TP4.disp_site$`Pr(>F)`[1], TP4.disp_os$`Pr(>F)`[1])

```


#### Effect Size
Calculate Effect Size of Origin for each Site
```{r}
##KL
##PERMANOVA
Coral.TP4.perm_KL<-adonis2(vegdist(Coral_TP4[which(Coral_TP4$Site=="KL"),c(11:13)], "euclidean")~ Coral_TP4$Origin[which(Coral_TP4$Site=="KL")], data=Coral_TP4[which(Coral_TP4$Site=="KL"),], strata=Coral_TP4$Genotype[which(Coral_TP4$Site=="KL")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP4.perm_KL)

##Check dispersion by Origin
TP4.KL.disp_orig<-anova(betadisper(vegdist(Coral_TP4[which(Coral_TP4$Site=="KL"),c(11:13)], "euclidean"), Coral_TP4$Origin[which(Coral_TP4$Site=="KL")]))
TP4.KL.disp_orig

##SS
##PERMANOVA
Coral.TP4.perm_SS<-adonis2(vegdist(Coral_TP4[which(Coral_TP4$Site=="SS"),c(11:13)], "euclidean")~ Coral_TP4$Origin[which(Coral_TP4$Site=="SS")], data=Coral_TP4[which(Coral_TP4$Site=="SS"),], strata=Coral_TP4$Genotype[which(Coral_TP4$Site=="SS")], method="euclidean")

##Effect Size
adonis_OmegaSq(Coral.TP4.perm_SS)

##Check dispersion by Origin
TP4.SS.disp_orig<-anova(betadisper(vegdist(Coral_TP4[which(Coral_TP4$Site=="SS"),c(11:13)], "euclidean"), Coral_TP4$Origin[which(Coral_TP4$Site=="SS")]))
TP4.SS.disp_orig

##Save results
PERM_T4.site.res<-data.frame(adonis_OmegaSq(Coral.TP4.perm_KL)[1,])
PERM_T4.site.res<-rbind(PERM_T4.site.res, data.frame(adonis_OmegaSq(Coral.TP4.perm_SS)[1,]))
PERM_T4.site.res$Predictor<-c("KL Origin", "SS Origin")
PERM_T4.site.res$p_DISP<-c(TP4.KL.disp_orig$`Pr(>F)`[1], TP4.SS.disp_orig$`Pr(>F)`[1])

##Combine results 
PERM_T4.res<-rbind(PERM_T4.res, PERM_T4.site.res)
PERM_T4.res$Timepoint<-rep("TP4", nrow(PERM_T4.res))

```


# Univariate 
Similar to the Multivariate analysis, checking for an overall effect of Origin or Site or Time, along with all interactions. Overall effect of Time alone (season) is not of main interest, but is predicted to interact with the effect of Origin as the time post-transplant increases.  

### Protein

#### Model
```{r}
##Check normality
hist(Coral$TP_ug.cm2)
shapiro.test(Coral$TP_ug.cm2)
#Normal

##Model 
#Function of Site and Origin and Time, with Genotype as a Random effect
#All interactions included
Prot.lme<-lmer(TP_ug.cm2~Origin*Site*TimeP+(1|Genotype), data=Coral)

##Check residuals
Prot.lme_res <- simulateResiduals(fittedModel = Prot.lme, plot = F)
plot(Prot.lme_res)

##Model results
anova(Prot.lme)

eta_squared(Prot.lme)

##Save model results
Prot.res<-data.frame(anova(Prot.lme))
Prot.res$Predictor<-c("Origin", "Site", "Time", "Origin x Site", "Origin x Time", "Site x Time", "Origin x Site x Time")
Prot.res$EtaSq<-c(eta_squared(Prot.lme)$Eta2)
Prot.res$Response<-rep("Protein", nrow(Prot.res))

```


#### Pairwise
```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site and Timepoint
Prot.pair<-emmeans(Prot.lme, pairwise~Origin | Site*TimeP)
Prot.pair

#Calculate standardized effect sizes for pairwise comparisons
Prot.pair.es<-data.frame(eff_size(Prot.pair, sigma=sigma(Prot.lme), edf=df.residual(Prot.lme)))
Prot.pair.es
Prot.pair.es<-Prot.pair.es %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)


##Save Results
Prot.pair.res<-merge(data.frame(Prot.pair$contrasts), Prot.pair.es[,-c(1)])
Prot.pair.res$Response<-rep("Protein", nrow(Prot.pair.res))

```


### Biomass

#### Model
```{r}
##Check normality
hist(Coral$AFDW_mg.cm2)
shapiro.test(Coral$AFDW_mg.cm2)
#Normal

##Model 
#Function of Site and Origin and Time, with Genotype as a Random effect
#All Interactions included
Bio.lme<-lmer(AFDW_mg.cm2~Origin*Site*TimeP+(1|Genotype), data=Coral)

##Check residuals
Bio.lme_res <- simulateResiduals(fittedModel = Bio.lme, plot = F)
plot(Bio.lme_res)

##Model results
anova(Bio.lme)

eta_squared(Bio.lme)

##Save model results
Bio.res<-data.frame(anova(Bio.lme))
Bio.res$Predictor<-c("Origin", "Site", "Time", "Origin x Site", "Origin x Time", "Site x Time", "Origin x Site x Time")
Bio.res$EtaSq<-c(eta_squared(Bio.lme)$Eta2)
Bio.res$Response<-rep("Biomass", nrow(Bio.res))

```


#### Pairwise
```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site and Timepoint
Bio.pair<-emmeans(Bio.lme, pairwise~Origin | Site*TimeP)
Bio.pair

#Calculate standardized effect sizes for pairwise comparisons
Bio.pair.es<-data.frame(eff_size(Bio.pair, sigma=sigma(Bio.lme), edf=df.residual(Bio.lme)))
Bio.pair.es
Bio.pair.es<-Bio.pair.es %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Bio.pair.res<-merge(data.frame(Bio.pair$contrasts), Bio.pair.es[,-c(1)])
Bio.pair.res$Response<-rep("Biomass", nrow(Bio.pair.res))
```


### Chlorophyll

#### Model
```{r}
##Check normality
hist(Coral$Chl_ug.cm2)
shapiro.test(Coral$Chl_ug.cm2)
#Not normal

hist(log(Coral$Chl_ug.cm2+1))
shapiro.test(log(Coral$Chl_ug.cm2+1))
#Improved

##Model 
#Function of Site and Origin and Time, with Genotype as a Random effect
#All Interactions included
Chl.lme<-lmer(log(Chl_ug.cm2+1)~Origin*Site*TimeP+(1|Genotype), data=Coral)

##Check residuals
Chl.lme_res <- simulateResiduals(fittedModel = Chl.lme, plot = F)
plot(Chl.lme_res)

##Model results
anova(Chl.lme)

eta_squared(Chl.lme)

##Save model results
Chl.res<-data.frame(anova(Chl.lme))
Chl.res$Predictor<-c("Origin", "Site", "Time", "Origin x Site", "Origin x Time", "Site x Time", "Origin x Site x Time")
Chl.res$EtaSq<-c(eta_squared(Chl.lme)$Eta2)
Chl.res$Response<-rep("Chlorophyll", nrow(Chl.res))

```


#### Pairwise
```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site and Timepoint
Chl.pair<-emmeans(Chl.lme, pairwise~Origin | Site*TimeP)
Chl.pair

#Calculate standardized effect sizes for pairwise comparisons
Chl.pair.es<-data.frame(eff_size(Chl.pair, sigma=sigma(Chl.lme), edf=df.residual(Chl.lme)))
Chl.pair.es
Chl.pair.es<-Chl.pair.es %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Chl.pair.res<-merge(data.frame(Chl.pair$contrasts), Chl.pair.es[,-c(1)])
Chl.pair.res$Response<-rep("Chlorophyll", nrow(Chl.pair.res))
```


### Save Results
```{r}
##Combine Results
Phys.res<-rbind(Prot.res, Bio.res, Chl.res)

##Add Timepoint
Phys.res$TimeP<-rep("All", nrow(Phys.res))

```


### Native v Transplant Effect Sizes
```{r}
##Combine Pairwise Results
Phys.pair.res<-rbind(Prot.pair.res, Bio.pair.res, Chl.pair.res)

##Add Significance levels
Phys.pair.res$Sig<-ifelse(Phys.pair.res$p.value<0.001, "***", ifelse(Phys.pair.res$p.value<0.01, "**", 
            ifelse(Phys.pair.res$p.value<0.05, "*", ifelse(Phys.pair.res$p.value<0.1, "-", NA))))

Phys.pair.res$Significant<-ifelse(Phys.pair.res$p.value<0.05, "sig", "non")

##Set Order
Phys.pair.res$Response<-factor(Phys.pair.res$Response, levels=c("Biomass", "Protein", "Chlorophyll"), ordered=TRUE)
```


#### Plot Effect Size of Origin
```{r Plot Origin Effect Sizes}
##Plot Effect Size across Timepoints
Phys.Orig.ES.plot<-ggplot(Phys.pair.res, aes(x=TimeP, y=effect.size)) + 
  geom_rect(fill="grey95", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.8, ymax=0.8)+
  geom_rect(fill="grey85", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.5, ymax=0.5)+
  geom_rect(fill="grey70", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.2, ymax=0.2)+
  geom_hline(yintercept=0, linetype="dashed", color="grey50", linewidth=0.5)+
  geom_errorbar(aes(ymin=effect.size-SE.es, ymax=effect.size+SE.es, colour=Site), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(aes(shape=Significant, colour=Site), size=point.sz, position=position_dodge(width=0.5))+
  facet_wrap(vars(Response))+
  scale_colour_manual(values=Site.colors.o)+
  scale_shape_manual(values=c(1, 16))+
  theme_bw()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position=c(.1, .1),
        legend.direction = "horizontal",
        strip.text=element_text(size=axis.title.sz),
        strip.background = element_rect(fill="white"), 
        panel.grid.major.y=element_blank())+
  labs(x="Time Point", y="Origin Effect Size (Cohen's d)", colour="Site")+
  guides(shape="none"); Phys.Orig.ES.plot
```


# Univariate by Timepoint
Given the interactions with Time, model each timepoint separately to compare the effect of Origin or current Site. Using partial Eta Squared as effect size within models with interactions.

### TP1

```{r}
##Subset Timepoint 1 
Coral.TP1<-subset(Coral, TimeP=="TP1")

```


#### Protein
```{r}
##Check normality
hist(Coral.TP1$TP_ug.cm2)
shapiro.test(Coral.TP1$TP_ug.cm2)
#Normal

##Model 
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Prot.lme_TP1<-lmer(TP_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP1)

##Check residuals
Prot.lme_res_TP1 <- simulateResiduals(fittedModel = Prot.lme_TP1, plot = F)
plot(Prot.lme_res_TP1)

##Model results
summary(Prot.lme_TP1)

eta_squared(Prot.lme_TP1)

##Save model results
Prot_TP1.res<-data.frame(summary(Prot.lme_TP1)$coefficients[-1,])
Prot_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Prot_TP1.res$EtaSq<-c(eta_squared(Prot.lme_TP1)$Eta2)
Prot_TP1.res$Response<-rep("Protein", nrow(Prot_TP1.res))
```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Prot.pair_TP1<-emmeans(Prot.lme_TP1, pairwise~Origin | Site)
Prot.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Prot.pair.es_TP1<-data.frame(eff_size(Prot.pair_TP1, sigma=sigma(Prot.lme_TP1), edf=df.residual(Prot.lme_TP1)))
Prot.pair.es_TP1
Prot.pair.es_TP1<-Prot.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Prot.pair_TP1.res<-merge(data.frame(Prot.pair_TP1$contrasts), Prot.pair.es_TP1[,-c(1)])
Prot.pair_TP1.res$Response<-rep("Protein", nrow(Prot.pair_TP1.res))

##Add Significance levels
Prot.pair_TP1.res$Sig<-ifelse(Prot.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Prot.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Prot.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Prot.pair_TP1.res$p.value<0.1, "-", NA))))
Prot.pair_TP1.res
```


```{r Plot Protein TP1}
##Summary statistics by Site and Origin
TP1_Prot.sum<-summarySE(Coral.TP1, measurevar="TP_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Protein across Treatments
TP1_Prot.plot<-ggplot(TP1_Prot.sum, aes(x=Site, y=TP_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TP_ug.cm2-se, ymax=TP_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Protein (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(550, 950)+
  annotate("text", x=2, y=950, label="**", size=sig.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=562, label.size=levels.sz, label="**", inherit.aes = FALSE); TP1_Prot.plot
```


#### Biomass 
```{r}
##Check normality
hist(Coral.TP1$AFDW_mg.cm2)
shapiro.test(Coral.TP1$AFDW_mg.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Bio.lme_TP1<-lmer(AFDW_mg.cm2~Origin*Site+(1|Genotype), data=Coral.TP1)

##Check residuals
Bio.lme_res_TP1 <- simulateResiduals(fittedModel = Bio.lme_TP1, plot = F)
plot(Bio.lme_res_TP1)

##Model results
summary(Bio.lme_TP1)

eta_squared(Bio.lme_TP1)

##Save model results
Bio_TP1.res<-data.frame(summary(Bio.lme_TP1)$coefficients[-1,])
Bio_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Bio_TP1.res$EtaSq<-c(eta_squared(Bio.lme_TP1)$Eta2)
Bio_TP1.res$Response<-rep("Biomass", nrow(Bio_TP1.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Bio.pair_TP1<-emmeans(Bio.lme_TP1, pairwise~Origin | Site)
Bio.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Bio.pair.es_TP1<-data.frame(eff_size(Bio.pair_TP1, sigma=sigma(Bio.lme_TP1), edf=df.residual(Bio.lme_TP1)))
Bio.pair.es_TP1
Bio.pair.es_TP1<-Bio.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Bio.pair_TP1.res<-merge(data.frame(Bio.pair_TP1$contrasts), Bio.pair.es_TP1[,-c(1)])
Bio.pair_TP1.res$Response<-rep("Biomass", nrow(Bio.pair_TP1.res))

##Add Significance levels
Bio.pair_TP1.res$Sig<-ifelse(Bio.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Bio.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Bio.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Bio.pair_TP1.res$p.value<0.1, "-", NA))))
Bio.pair_TP1.res
```


```{r Plot Biomass TP1}
##Summary statistics by Site and Origin
TP1_Bio.sum<-summarySE(Coral.TP1, measurevar="AFDW_mg.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Biomass across Treatments
TP1_Bio.plot<-ggplot(TP1_Bio.sum, aes(x=Site, y=AFDW_mg.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=AFDW_mg.cm2-se, ymax=AFDW_mg.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz),
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position = "top")+
  labs(x="Site and Origin", y=expression(paste('Biomass (mg cm'^-2*")")), colour=NULL)+
  ylim(1.35, 2.45)+
   annotate("text", x=2, y=2.1, label="*", size=sig.sz, fontface="bold"); TP1_Bio.plot
```


#### Chlorophyll
```{r}
##Check normality
hist(Coral.TP1$Chl_ug.cm2)
shapiro.test(Coral.TP1$Chl_ug.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Chl.lme_TP1<-lmer(Chl_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP1)

##Check residuals
Chl.lme_res_TP1 <- simulateResiduals(fittedModel = Chl.lme_TP1, plot = F)
plot(Chl.lme_res_TP1)

##Model results
summary(Chl.lme_TP1)

eta_squared(Chl.lme_TP1)

##Save model results
Chl_TP1.res<-data.frame(summary(Chl.lme_TP1)$coefficients[-1,])
Chl_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Chl_TP1.res$EtaSq<-c(eta_squared(Chl.lme_TP1)$Eta2)
Chl_TP1.res$Response<-rep("Chlorophyll", nrow(Chl_TP1.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Chl.pair_TP1<-emmeans(Chl.lme_TP1, pairwise~Origin | Site)
Chl.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Chl.pair.es_TP1<-data.frame(eff_size(Chl.pair_TP1, sigma=sigma(Chl.lme_TP1), edf=df.residual(Chl.lme_TP1)))
Chl.pair.es_TP1
Chl.pair.es_TP1<-Chl.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Chl.pair_TP1.res<-merge(data.frame(Chl.pair_TP1$contrasts), Chl.pair.es_TP1[,-c(1)])
Chl.pair_TP1.res$Response<-rep("Chlorophyll", nrow(Chl.pair_TP1.res))

##Add Significance levels
Chl.pair_TP1.res$Sig<-ifelse(Chl.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Chl.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Chl.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Chl.pair_TP1.res$p.value<0.1, "-", NA))))
Chl.pair_TP1.res
```


```{r Plot Chlorophyll TP1}
##Summary statistics by Site and Origin
TP1_Chl.sum<-summarySE(Coral.TP1, measurevar="Chl_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Chlorophyll across Treatments
TP1_Chl.plot<-ggplot(TP1_Chl.sum, aes(x=Site, y=Chl_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl_ug.cm2-se, ymax=Chl_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Total Chlorophyll (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(0.5, 3.25)+
  annotate("text", x=2, y=1.7, label="***", size=sig.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=0.58, label.size=levels.sz, label="***", inherit.aes = FALSE); TP1_Chl.plot
```


#### Save Results
```{r}
##Combine Results
Phys_TP1.res<-rbind(Prot_TP1.res, Bio_TP1.res, Chl_TP1.res)
Phys.pair_TP1.res<-rbind(Prot.pair_TP1.res, Bio.pair_TP1.res, Chl.pair_TP1.res)

##Add Timepoint
Phys_TP1.res$TimeP<-rep("TP1", nrow(Phys_TP1.res))
Phys.pair_TP1.res$TimeP<-rep("TP1", nrow(Phys.pair_TP1.res))

```


### TP2

```{r}
##Subset Timepoint 2 
Coral.TP2<-subset(Coral, TimeP=="TP2")

```


#### Protein 
```{r}
##Check normality
hist(Coral.TP2$TP_ug.cm2)
shapiro.test(Coral.TP2$TP_ug.cm2)
#Normal

##Model 
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Prot.lme_TP2<-lmer(TP_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP2)

##Check residuals
Prot.lme_res_TP2 <- simulateResiduals(fittedModel = Prot.lme_TP2, plot = F)
plot(Prot.lme_res_TP2)

##Model results
summary(Prot.lme_TP2)

eta_squared(Prot.lme_TP2)

##Save model results
Prot_TP2.res<-data.frame(summary(Prot.lme_TP2)$coefficients[-1,])
Prot_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Prot_TP2.res$EtaSq<-c(eta_squared(Prot.lme_TP2)$Eta2)
Prot_TP2.res$Response<-rep("Protein", nrow(Prot_TP2.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Prot.pair_TP2<-emmeans(Prot.lme_TP2, pairwise~Origin | Site)
Prot.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Prot.pair.es_TP2<-data.frame(eff_size(Prot.pair_TP2, sigma=sigma(Prot.lme_TP2), edf=df.residual(Prot.lme_TP2)))
Prot.pair.es_TP2
Prot.pair.es_TP2<-Prot.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Prot.pair_TP2.res<-merge(data.frame(Prot.pair_TP2$contrasts), Prot.pair.es_TP2[,-c(1)])
Prot.pair_TP2.res$Response<-rep("Protein", nrow(Prot.pair_TP2.res))

##Add Significance levels
Prot.pair_TP2.res$Sig<-ifelse(Prot.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Prot.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Prot.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Prot.pair_TP2.res$p.value<0.1, "-", NA))))
Prot.pair_TP2.res
```



```{r Plot Protein TP2}
##Summary statistics by Site and Origin
TP2_Prot.sum<-summarySE(Coral.TP2, measurevar="TP_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Protein across Treatments
TP2_Prot.plot<-ggplot(TP2_Prot.sum, aes(x=Site, y=TP_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TP_ug.cm2-se, ymax=TP_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Protein (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(550, 950)+
  geom_bracket(xmin=1, xmax=2, y.position=562, label.size=levels.sz, label="*", inherit.aes = FALSE); TP2_Prot.plot
```


#### Biomass 
```{r}
##Check normality
hist(Coral.TP2$AFDW_mg.cm2)
shapiro.test(Coral.TP2$AFDW_mg.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Bio.lme_TP2<-lmer(AFDW_mg.cm2~Origin*Site+(1|Genotype), data=Coral.TP2)

##Check residuals
Bio.lme_res_TP2 <- simulateResiduals(fittedModel = Bio.lme_TP2, plot = F)
plot(Bio.lme_res_TP2)

##Model results
summary(Bio.lme_TP2)

eta_squared(Bio.lme_TP2)

##Save model results
Bio_TP2.res<-data.frame(summary(Bio.lme_TP2)$coefficients[-1,])
Bio_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Bio_TP2.res$EtaSq<-c(eta_squared(Bio.lme_TP2)$Eta2)
Bio_TP2.res$Response<-rep("Biomass", nrow(Bio_TP2.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Bio.pair_TP2<-emmeans(Bio.lme_TP2, pairwise~Origin | Site)
Bio.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Bio.pair.es_TP2<-data.frame(eff_size(Bio.pair_TP2, sigma=sigma(Bio.lme_TP2), edf=df.residual(Bio.lme_TP2)))
Bio.pair.es_TP2
Bio.pair.es_TP2<-Bio.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Bio.pair_TP2.res<-merge(data.frame(Bio.pair_TP2$contrasts), Bio.pair.es_TP2[,-c(1)])
Bio.pair_TP2.res$Response<-rep("Biomass", nrow(Bio.pair_TP2.res))

##Add Significance levels
Bio.pair_TP2.res$Sig<-ifelse(Bio.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Bio.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Bio.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Bio.pair_TP2.res$p.value<0.1, "-", NA))))
Bio.pair_TP2.res
```


```{r Plot Biomass TP2}
##Summary statistics by Site and Origin
TP2_Bio.sum<-summarySE(Coral.TP2, measurevar="AFDW_mg.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Biomass across Treatments
TP2_Bio.plot<-ggplot(TP2_Bio.sum, aes(x=Site, y=AFDW_mg.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=AFDW_mg.cm2-se, ymax=AFDW_mg.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Biomass (mg cm'^-2*")")), colour=NULL)+
  ylim(1.35, 2.45)+
  annotate("text", x=1, y=2, label="*", size=levels.sz, fontface="bold"); TP2_Bio.plot
```


#### Chlorophyll
```{r}
##Check normality
hist(Coral.TP2$Chl_ug.cm2)
shapiro.test(Coral.TP2$Chl_ug.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Chl.lme_TP2<-lmer(Chl_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP2)

##Check residuals
Chl.lme_res_TP2 <- simulateResiduals(fittedModel = Chl.lme_TP2, plot = F)
plot(Chl.lme_res_TP2)

##Model results
summary(Chl.lme_TP2)

eta_squared(Chl.lme_TP2)

##Save model results
Chl_TP2.res<-data.frame(summary(Chl.lme_TP2)$coefficients[-1,])
Chl_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Chl_TP2.res$EtaSq<-c(eta_squared(Chl.lme_TP2)$Eta2)
Chl_TP2.res$Response<-rep("Chlorophyll", nrow(Chl_TP2.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Chl.pair_TP2<-emmeans(Chl.lme_TP2, pairwise~Origin | Site)
Chl.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Chl.pair.es_TP2<-data.frame(eff_size(Chl.pair_TP2, sigma=sigma(Chl.lme_TP2), edf=df.residual(Chl.lme_TP2)))
Chl.pair.es_TP2
Chl.pair.es_TP2<-Chl.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Chl.pair_TP2.res<-merge(data.frame(Chl.pair_TP2$contrasts), Chl.pair.es_TP2[,-c(1)])
Chl.pair_TP2.res$Response<-rep("Chlorophyll", nrow(Chl.pair_TP2.res))

##Add Significance levels
Chl.pair_TP2.res$Sig<-ifelse(Chl.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Chl.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Chl.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Chl.pair_TP2.res$p.value<0.1, "-", NA))))
Chl.pair_TP2.res
```


```{r Plot Chlorophyll TP2}
##Summary statistics by Site and Origin
TP2_Chl.sum<-summarySE(Coral.TP2, measurevar="Chl_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Chlorophyll across Treatments
TP2_Chl.plot<-ggplot(TP2_Chl.sum, aes(x=Site, y=Chl_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl_ug.cm2-se, ymax=Chl_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position= "top")+
  labs(x="Site and Origin", y=expression(paste('Total Chlorophyll (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(0.5, 3.25)+
  annotate("text", x=2, y=2.6, label="*", size=sig.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=0.58, label.size=levels.sz, label="**", inherit.aes = FALSE); TP2_Chl.plot
```


#### Save Results
```{r}
##Combine Results
Phys_TP2.res<-rbind(Prot_TP2.res, Bio_TP2.res, Chl_TP2.res)
Phys.pair_TP2.res<-rbind(Prot.pair_TP2.res, Bio.pair_TP2.res, Chl.pair_TP2.res)

##Add Timepoint
Phys_TP2.res$TimeP<-rep("TP2", nrow(Phys_TP2.res))
Phys.pair_TP2.res$TimeP<-rep("TP2", nrow(Phys.pair_TP2.res))

```


### TP3

```{r}
##Subset Timepoint 3 
Coral.TP3<-subset(Coral, TimeP=="TP3")

```


#### Protein 
```{r}
##Check normality
hist(Coral.TP3$TP_ug.cm2)
shapiro.test(Coral.TP3$TP_ug.cm2)
#Normal

##Model 
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Prot.lme_TP3<-lmer(TP_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP3)

##Check residuals
Prot.lme_res_TP3 <- simulateResiduals(fittedModel = Prot.lme_TP3, plot = F)
plot(Prot.lme_res_TP3)

##Model results
summary(Prot.lme_TP3)

eta_squared(Prot.lme_TP3)

##Save model results
Prot_TP3.res<-data.frame(summary(Prot.lme_TP3)$coefficients[-1,])
Prot_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Prot_TP3.res$EtaSq<-c(eta_squared(Prot.lme_TP3)$Eta2)
Prot_TP3.res$Response<-rep("Protein", nrow(Prot_TP3.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Prot.pair_TP3<-emmeans(Prot.lme_TP3, pairwise~Origin | Site)
Prot.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Prot.pair.es_TP3<-data.frame(eff_size(Prot.pair_TP3, sigma=sigma(Prot.lme_TP3), edf=df.residual(Prot.lme_TP3)))
Prot.pair.es_TP3
Prot.pair.es_TP3<-Prot.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Prot.pair_TP3.res<-merge(data.frame(Prot.pair_TP3$contrasts), Prot.pair.es_TP3[,-c(1)])
Prot.pair_TP3.res$Response<-rep("Protein", nrow(Prot.pair_TP3.res))

##Add Significance levels
Prot.pair_TP3.res$Sig<-ifelse(Prot.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Prot.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Prot.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Prot.pair_TP3.res$p.value<0.1, "-", NA))))
Prot.pair_TP3.res
```



```{r Plot Protein TP3}
##Summary statistics by Site and Origin
TP3_Prot.sum<-summarySE(Coral.TP3, measurevar="TP_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Protein across Treatments
TP3_Prot.plot<-ggplot(TP3_Prot.sum, aes(x=Site, y=TP_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TP_ug.cm2-se, ymax=TP_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Protein (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(550, 950); TP3_Prot.plot
```


#### Biomass 
```{r}
##Check normality
hist(Coral.TP3$AFDW_mg.cm2)
shapiro.test(Coral.TP3$AFDW_mg.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Bio.lme_TP3<-lmer(AFDW_mg.cm2~Origin*Site+(1|Genotype), data=Coral.TP3)

##Check residuals
Bio.lme_res_TP3 <- simulateResiduals(fittedModel = Bio.lme_TP3, plot = F)
plot(Bio.lme_res_TP3)

##Model results
summary(Bio.lme_TP3)

eta_squared(Bio.lme_TP3)

##Save model results
Bio_TP3.res<-data.frame(summary(Bio.lme_TP3)$coefficients[-1,])
Bio_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Bio_TP3.res$EtaSq<-c(eta_squared(Bio.lme_TP3)$Eta2)
Bio_TP3.res$Response<-rep("Biomass", nrow(Bio_TP3.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Bio.pair_TP3<-emmeans(Bio.lme_TP3, pairwise~Origin | Site)
Bio.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Bio.pair.es_TP3<-data.frame(eff_size(Bio.pair_TP3, sigma=sigma(Bio.lme_TP3), edf=df.residual(Bio.lme_TP3)))
Bio.pair.es_TP3
Bio.pair.es_TP3<-Bio.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Bio.pair_TP3.res<-merge(data.frame(Bio.pair_TP3$contrasts), Bio.pair.es_TP3[,-c(1)])
Bio.pair_TP3.res$Response<-rep("Biomass", nrow(Bio.pair_TP3.res))

##Add Significance levels
Bio.pair_TP3.res$Sig<-ifelse(Bio.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Bio.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Bio.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Bio.pair_TP3.res$p.value<0.1, "-", NA))))
Bio.pair_TP3.res
```


```{r Plot Biomass TP3}
##Summary statistics by Site and Origin
TP3_Bio.sum<-summarySE(Coral.TP3, measurevar="AFDW_mg.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Biomass across Treatments
TP3_Bio.plot<-ggplot(TP3_Bio.sum, aes(x=Site, y=AFDW_mg.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=AFDW_mg.cm2-se, ymax=AFDW_mg.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Biomass (mg cm'^-2*")")), colour=NULL)+
  ylim(1.35, 2.45); TP3_Bio.plot
```


#### Chlorophyll
```{r}
##Check normality
hist(Coral.TP3$Chl_ug.cm2)
shapiro.test(Coral.TP3$Chl_ug.cm2)
#Not normal

hist(log(Coral.TP3$Chl_ug.cm2+1))
shapiro.test(log(Coral.TP3$Chl_ug.cm2+1))
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Chl.lme_TP3<-lmer(log(Chl_ug.cm2+1)~Origin*Site+(1|Genotype), data=Coral.TP3)

##Check residuals
Chl.lme_res_TP3 <- simulateResiduals(fittedModel = Chl.lme_TP3, plot = F)
plot(Chl.lme_res_TP3)

##Model results
summary(Chl.lme_TP3)

eta_squared(Chl.lme_TP3)

##Save model results
Chl_TP3.res<-data.frame(summary(Chl.lme_TP3)$coefficients[-1,])
Chl_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Chl_TP3.res$EtaSq<-c(eta_squared(Chl.lme_TP3)$Eta2)
Chl_TP3.res$Response<-rep("Chlorophyll", nrow(Chl_TP3.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Chl.pair_TP3<-emmeans(Chl.lme_TP3, pairwise~Origin | Site)
Chl.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Chl.pair.es_TP3<-data.frame(eff_size(Chl.pair_TP3, sigma=sigma(Chl.lme_TP3), edf=df.residual(Chl.lme_TP3)))
Chl.pair.es_TP3
Chl.pair.es_TP3<-Chl.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Chl.pair_TP3.res<-merge(data.frame(Chl.pair_TP3$contrasts), Chl.pair.es_TP3[,-c(1)])
Chl.pair_TP3.res$Response<-rep("Chlorophyll", nrow(Chl.pair_TP3.res))

##Add Significance levels
Chl.pair_TP3.res$Sig<-ifelse(Chl.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Chl.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Chl.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Chl.pair_TP3.res$p.value<0.1, "-", NA))))
Chl.pair_TP3.res
```


```{r Plot Chlorophyll TP3}
##Summary statistics by Site and Origin
TP3_Chl.sum<-summarySE(Coral.TP3, measurevar="Chl_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Chlorophyll across Treatments
TP3_Chl.plot<-ggplot(TP3_Chl.sum, aes(x=Site, y=Chl_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl_ug.cm2-se, ymax=Chl_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Total Chlorophyll (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(0.5, 3.25)+
  annotate("text", x=1, y=1.5, label="-", size=levels.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=0.58, label.size=levels.sz, label="**", inherit.aes = FALSE); TP3_Chl.plot
```


#### Save Results
```{r}
##Combine Results
Phys_TP3.res<-rbind(Prot_TP3.res, Bio_TP3.res, Chl_TP3.res)
Phys.pair_TP3.res<-rbind(Prot.pair_TP3.res, Bio.pair_TP3.res, Chl.pair_TP3.res)

##Add Timepoint
Phys_TP3.res$TimeP<-rep("TP3", nrow(Phys_TP3.res))
Phys.pair_TP3.res$TimeP<-rep("TP3", nrow(Phys.pair_TP3.res))

```


### TP4

```{r}
##Subset Timepoint 4 
Coral.TP4<-subset(Coral, TimeP=="TP4")

```


#### Protein 
```{r}
##Check normality
hist(Coral.TP4$TP_ug.cm2)
shapiro.test(Coral.TP4$TP_ug.cm2)
#Normal

##Model 
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Prot.lme_TP4<-lmer(TP_ug.cm2~Origin*Site+(1|Genotype), data=Coral.TP4)

##Check residuals
Prot.lme_res_TP4 <- simulateResiduals(fittedModel = Prot.lme_TP4, plot = F)
plot(Prot.lme_res_TP4)

##Model results
summary(Prot.lme_TP4)

eta_squared(Prot.lme_TP4)

##Save model results
Prot_TP4.res<-data.frame(summary(Prot.lme_TP4)$coefficients[-1,])
Prot_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Prot_TP4.res$EtaSq<-c(eta_squared(Prot.lme_TP4)$Eta2)
Prot_TP4.res$Response<-rep("Protein", nrow(Prot_TP4.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Prot.pair_TP4<-emmeans(Prot.lme_TP4, pairwise~Origin | Site)
Prot.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Prot.pair.es_TP4<-data.frame(eff_size(Prot.pair_TP4, sigma=sigma(Prot.lme_TP4), edf=df.residual(Prot.lme_TP4)))
Prot.pair.es_TP4
Prot.pair.es_TP4<-Prot.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Prot.pair_TP4.res<-merge(data.frame(Prot.pair_TP4$contrasts), Prot.pair.es_TP4[,-c(1)])
Prot.pair_TP4.res$Response<-rep("Protein", nrow(Prot.pair_TP4.res))

##Add Significance levels
Prot.pair_TP4.res$Sig<-ifelse(Prot.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Prot.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Prot.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Prot.pair_TP4.res$p.value<0.1, "-", NA))))
Prot.pair_TP4.res
```



```{r Plot Protein TP4}
##Summary statistics by Site and Origin
TP4_Prot.sum<-summarySE(Coral.TP4, measurevar="TP_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Protein across Treatments
TP4_Prot.plot<-ggplot(TP4_Prot.sum, aes(x=Site, y=TP_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TP_ug.cm2-se, ymax=TP_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Protein (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(550, 950)+
  geom_bracket(xmin=1, xmax=2, y.position=562, label.size=levels.sz, label="**", inherit.aes = FALSE); TP4_Prot.plot
```


#### Biomass 
```{r}
##Check normality
hist(Coral.TP4$AFDW_mg.cm2)
shapiro.test(Coral.TP4$AFDW_mg.cm2)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Bio.lme_TP4<-lmer(AFDW_mg.cm2~Origin*Site+(1|Genotype), data=Coral.TP4)

##Check residuals
Bio.lme_res_TP4 <- simulateResiduals(fittedModel = Bio.lme_TP4, plot = F)
plot(Bio.lme_res_TP4)

##Model results
summary(Bio.lme_TP4)

eta_squared(Bio.lme_TP4)

##Save model results
Bio_TP4.res<-data.frame(summary(Bio.lme_TP4)$coefficients[-1,])
Bio_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Bio_TP4.res$EtaSq<-c(eta_squared(Bio.lme_TP4)$Eta2)
Bio_TP4.res$Response<-rep("Biomass", nrow(Bio_TP4.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Bio.pair_TP4<-emmeans(Bio.lme_TP4, pairwise~Origin | Site)
Bio.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Bio.pair.es_TP4<-data.frame(eff_size(Bio.pair_TP4, sigma=sigma(Bio.lme_TP4), edf=df.residual(Bio.lme_TP4)))
Bio.pair.es_TP4
Bio.pair.es_TP4<-Bio.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Bio.pair_TP4.res<-merge(data.frame(Bio.pair_TP4$contrasts), Bio.pair.es_TP4[,-c(1)])
Bio.pair_TP4.res$Response<-rep("Biomass", nrow(Bio.pair_TP4.res))

##Add Significance levels
Bio.pair_TP4.res$Sig<-ifelse(Bio.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Bio.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Bio.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Bio.pair_TP4.res$p.value<0.1, "-", NA))))
Bio.pair_TP4.res
```


```{r Plot Biomass TP4}
##Summary statistics by Site and Origin
TP4_Bio.sum<-summarySE(Coral.TP4, measurevar="AFDW_mg.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Biomass across Treatments
TP4_Bio.plot<-ggplot(TP4_Bio.sum, aes(x=Site, y=AFDW_mg.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=AFDW_mg.cm2-se, ymax=AFDW_mg.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Biomass (mg cm'^-2*")")), colour=NULL)+
  ylim(1.35, 2.45)+
  geom_bracket(xmin=1, xmax=2, y.position=1.38, label.size=levels.sz, label="**", inherit.aes = FALSE); TP4_Bio.plot
```


#### Chlorophyll
```{r}
##Check normality
hist(Coral.TP4$Chl_ug.cm2)
shapiro.test(Coral.TP4$Chl_ug.cm2)
#Slightly non normal

hist(log(Coral.TP4$Chl_ug.cm2+1))
shapiro.test(log(Coral.TP4$Chl_ug.cm2+1))
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interactions between Origin and Site
Chl.lme_TP4<-lmer(log(Chl_ug.cm2+1)~Origin*Site+(1|Genotype), data=Coral.TP4)

##Check residuals
Chl.lme_res_TP4 <- simulateResiduals(fittedModel = Chl.lme_TP4, plot = F)
plot(Chl.lme_res_TP4)

##Model results
summary(Chl.lme_TP4)

eta_squared(Chl.lme_TP4)

##Save model results
Chl_TP4.res<-data.frame(summary(Chl.lme_TP4)$coefficients[-1,])
Chl_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Chl_TP4.res$EtaSq<-c(eta_squared(Chl.lme_TP4)$Eta2)
Chl_TP4.res$Response<-rep("Chlorophyll", nrow(Chl_TP4.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Chl.pair_TP4<-emmeans(Chl.lme_TP4, pairwise~Origin | Site)
Chl.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Chl.pair.es_TP4<-data.frame(eff_size(Chl.pair_TP4, sigma=sigma(Chl.lme_TP4), edf=df.residual(Chl.lme_TP4)))
Chl.pair.es_TP4
Chl.pair.es_TP4<-Chl.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Chl.pair_TP4.res<-merge(data.frame(Chl.pair_TP4$contrasts), Chl.pair.es_TP4[,-c(1)])
Chl.pair_TP4.res$Response<-rep("Chlorophyll", nrow(Chl.pair_TP4.res))

##Add Significance levels
Chl.pair_TP4.res$Sig<-ifelse(Chl.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Chl.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Chl.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Chl.pair_TP4.res$p.value<0.1, "-", NA))))
Chl.pair_TP4.res
```


```{r Plot Chlorophyll TP4}
##Summary statistics by Site and Origin
TP4_Chl.sum<-summarySE(Coral.TP4, measurevar="Chl_ug.cm2", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Chlorophyll across Treatments
TP4_Chl.plot<-ggplot(TP4_Chl.sum, aes(x=Site, y=Chl_ug.cm2, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl_ug.cm2-se, ymax=Chl_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position="top")+
  labs(x="Site and Origin", y=expression(paste('Total Chlorophyll (\u03BCg cm'^-2*")")), colour=NULL)+
  ylim(0.5, 3.25)+
  geom_bracket(xmin=1, xmax=2, y.position=0.58, label.size=levels.sz, label="***", inherit.aes = FALSE); TP4_Chl.plot
```


#### Save Results
```{r}
##Combine Results
Phys_TP4.res<-rbind(Prot_TP4.res, Bio_TP4.res, Chl_TP4.res)
Phys.pair_TP4.res<-rbind(Prot.pair_TP4.res, Bio.pair_TP4.res, Chl.pair_TP4.res)

##Add Timepoint
Phys_TP4.res$TimeP<-rep("TP4", nrow(Phys_TP4.res))
Phys.pair_TP4.res$TimeP<-rep("TP4", nrow(Phys.pair_TP4.res))

```


### Current v Original Site Effect Sizes
```{r}
##Combine Results
Phys_TP.res<-rbind(Phys_TP1.res, Phys_TP2.res, Phys_TP3.res, Phys_TP4.res)
Phys_TP.res<-Phys_TP.res %>% dplyr::rename(p = Pr...t..)

##Add Significance levels
Phys_TP.res$Sig<-ifelse(Phys_TP.res$p<0.001, "***", ifelse(Phys_TP.res$p<0.01, "**", 
            ifelse(Phys_TP.res$p<0.05, "*", ifelse(Phys_TP.res$p<0.1, "-", NA))))

##Set Order
Phys_TP.res$Response<-factor(Phys_TP.res$Response, levels=c("Biomass", "Protein", "Chlorophyll"), ordered=TRUE)
```


#### Plot Effect Size of Origin and Site
```{r Plot Origin and Site Effect Sizes}
Phys.Site.Orig.ES.plot<-ggplot(Phys_TP.res, aes(fill=Predictor, y=EtaSq, x=TimeP)) + 
    geom_bar(position="stack", stat="identity")+
  facet_wrap(vars(Response))+
  theme_bw()+
  scale_fill_manual(values=c("indianred2", "mediumpurple3", "steelblue2"))+
  theme(strip.text=element_text(size=axis.title.sz),
        strip.background = element_rect(fill="white"),
        axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position=c(.08, .8), 
        panel.grid.major.x=element_blank())+
   labs(x="Time Point", y=expression(paste("Predictor Effect Size (p", eta^2, ")")), colour="Predictor")+
   geom_text(aes(label=Sig), position=position_stack(vjust=0.5), size=sig.sz); Phys.Site.Orig.ES.plot
```


# Figures

### Figure 3 Univariate Physiology Effect Sizes

#### Adjust for Panel
```{r}
Phys.Site.Orig.ES.plot<-Phys.Site.Orig.ES.plot+ labs(x="")
```


#### Figure 3
```{r}
##Create Panel
Phys.ES_fig<-plot_grid(Phys.Site.Orig.ES.plot, Phys.Orig.ES.plot, 
                    rel_widths=1, rel_heights = 1,
                    nrow=2, ncol=1, byrow=T, labels = c("A", "B"), 
                    label_size=panel.lab.sz)

##Save Figure
ggsave(filename="Figures/02_Physiology/Fig3_PhysiologyEffectSizes.png", plot=Phys.ES_fig, dpi=300, width=12, height=9, units="in")

```

### Figure S4 Univariate Physiology Metrics

#### Adjust for Panel
```{r}
##Biomass
TP1_Bio.plot<-TP1_Bio.plot+ ggtitle("TP1")+ labs(x=NULL)+
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position=c(0.5, 0.85), legend.direction="horizontal")+
  guides(color=guide_legend(nrow=2, byrow=FALSE)) 

TP2_Bio.plot<-TP2_Bio.plot+ ggtitle("TP2")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")

TP3_Bio.plot<-TP3_Bio.plot+ ggtitle("TP3")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")

TP4_Bio.plot<-TP4_Bio.plot+ ggtitle("TP4")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")

##Protein
TP1_Prot.plot<-TP1_Prot.plot + labs(x=NULL) + theme(legend.position="none")

TP2_Prot.plot<-TP2_Prot.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")

TP3_Prot.plot<-TP3_Prot.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")

TP4_Prot.plot<-TP4_Prot.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")

##Chlorophyll
TP1_Chl.plot<-TP1_Chl.plot + theme(legend.position="none")

TP2_Chl.plot<-TP2_Chl.plot + labs(y=NULL) + theme(legend.position="none")

TP3_Chl.plot<-TP3_Chl.plot + labs(y=NULL) + theme(legend.position="none")

TP4_Chl.plot<-TP4_Chl.plot + labs(y=NULL) + theme(legend.position="none")
```


#### Figure S4
```{r}
##Create Panel
Phys_fig<-plot_grid(TP1_Bio.plot, TP2_Bio.plot, 
                      TP3_Bio.plot, TP4_Bio.plot,  
                      TP1_Prot.plot, TP2_Prot.plot, 
                      TP3_Prot.plot, TP4_Prot.plot,  
                      TP1_Chl.plot, TP2_Chl.plot, 
                      TP3_Chl.plot, TP4_Chl.plot,
                    rel_widths=1, rel_heights = 1,
                    nrow=3, ncol=4, byrow=T, labels = NULL)

##Save Figure
ggsave(filename="Figures/02_Physiology/FigS4_PhysiologyMetrics.png", plot=Phys_fig, dpi=300, width=14, height=12, units="in")

```




# Tables 

### Table S3A Physiology PERMANOVA and PERMDISP
```{r}
##Combine Results Tables
TableS3A_Phys.PERM<-data.frame(rbind(PERM.res, PERM_T1.res, PERM_T2.res, PERM_T3.res, PERM_T4.res))

##Organize
names(TableS3A_Phys.PERM)
TableS3A_Phys.PERM<-TableS3A_Phys.PERM %>% dplyr::rename( DF = Df, SS = SumOfSqs, EffectSize = parOmegaSq, p = Pr..F.)
TableS3A_Phys.PERM<-TableS3A_Phys.PERM[,c("Timepoint", "Predictor", "DF", "SS", "F",   "EffectSize", "p", "p_DISP")]

#Round to 3 digits
TableS3A_Phys.PERM$SS<-round(TableS3A_Phys.PERM$SS, 3)
TableS3A_Phys.PERM$F<-round(TableS3A_Phys.PERM$F, 3)
TableS3A_Phys.PERM$EffectSize<-round(TableS3A_Phys.PERM$EffectSize, 3)
TableS3A_Phys.PERM$p<-round(TableS3A_Phys.PERM$p, 3)
TableS3A_Phys.PERM$p_DISP<-round(TableS3A_Phys.PERM$p_DISP, 3)

##Write Out Table
write.csv(TableS3A_Phys.PERM, "Tables/TableS3A_Physiology_PERM_Results.csv", row.names=FALSE)


```


### Table S4A Physiology Univariate LM all Timepoints
```{r}
##Organize
TableS4A_Phys.LM_All<-Phys.res
names(TableS4A_Phys.LM_All)
TableS4A_Phys.LM_All<-TableS4A_Phys.LM_All %>% dplyr::rename(SumSq = Sum.Sq, MeanSq= Mean.Sq, F = F.value, p = Pr..F., Timepoint = TimeP)
TableS4A_Phys.LM_All<-TableS4A_Phys.LM_All[,c("Timepoint", "Response", "Predictor", "SumSq", "MeanSq", "NumDF", "DenDF", "F", "p", "EtaSq")]

#Round to 3 digits
TableS4A_Phys.LM_All$SumSq<-round(TableS4A_Phys.LM_All$SumSq, 3)
TableS4A_Phys.LM_All$MeanSq<-round(TableS4A_Phys.LM_All$MeanSq, 3)
TableS4A_Phys.LM_All$F<-round(TableS4A_Phys.LM_All$F, 3)
TableS4A_Phys.LM_All$p<-round(TableS4A_Phys.LM_All$p, 3)
TableS4A_Phys.LM_All$EtaSq<-round(TableS4A_Phys.LM_All$EtaSq, 3)

#Integer
TableS4A_Phys.LM_All$DenDF<-round(TableS4A_Phys.LM_All$DenDF, 0)

##Write Out Table
write.csv(TableS4A_Phys.LM_All, "Tables/TableS4A_Physiology_LM_Results_AllTP.csv", row.names=FALSE)

```


### Table S4B Physiology Univariate LM by Timepoint

```{r}
##Organize
TableS4B_Phys.LM_TP<-Phys_TP.res
names(TableS4B_Phys.LM_TP)
TableS4B_Phys.LM_TP<-TableS4B_Phys.LM_TP %>% dplyr::rename(SE = Std..Error, DF = df, t = t.value, Timepoint = TimeP)
TableS4B_Phys.LM_TP<-TableS4B_Phys.LM_TP[,c("Timepoint", "Response", "Predictor", "Estimate", "SE", "DF", "t", "p", "EtaSq")]

#Round to 3 digits
TableS4B_Phys.LM_TP$Estimate<-round(TableS4B_Phys.LM_TP$Estimate, 3)
TableS4B_Phys.LM_TP$SE<-round(TableS4B_Phys.LM_TP$SE, 3)
TableS4B_Phys.LM_TP$t<-round(TableS4B_Phys.LM_TP$t, 3)
TableS4B_Phys.LM_TP$p<-round(TableS4B_Phys.LM_TP$p, 3)
TableS4B_Phys.LM_TP$EtaSq<-round(TableS4B_Phys.LM_TP$EtaSq, 3)

#Integer
TableS4B_Phys.LM_TP$DF<-round(TableS4B_Phys.LM_TP$DF, 0)

##Write Out Table
write.csv(TableS4B_Phys.LM_TP, "Tables/TableS4B_Physiology_LM_Results_TP.csv", row.names=FALSE)

```


### Table S5 Physiology Univariate Pairwise
```{r}
##Combine Results of TP Models
Phys.pair.res_TP<-rbind(Phys.pair_TP1.res, Phys.pair_TP2.res, Phys.pair_TP3.res, Phys.pair_TP4.res)

##Rename p.value for merging
Phys.pair.res_TP<-Phys.pair.res_TP %>% dplyr::rename(p_Timepoint = p.value)

##Combine Results Tables
Phys.pair.res.all<-merge(Phys.pair.res, Phys.pair.res_TP[,c("Site", "TimeP", "Response", "p_Timepoint")])
TableS5_Phys.Pair<-Phys.pair.res.all

##Organize
names(TableS5_Phys.Pair)
TableS5_Phys.Pair<-TableS5_Phys.Pair %>% dplyr::rename(Timepoint = TimeP, Contrast =contrast, Estimate = estimate, DF = df, t = t.ratio, p = p.value, EffectSize = effect.size, EffectSize.SE = SE.es)
TableS5_Phys.Pair<-TableS5_Phys.Pair[,c("Timepoint", "Response", "Site", "Contrast", "Estimate", "SE", "DF", "t", "p", "p_Timepoint", "EffectSize", "EffectSize.SE")]

#Round to 3 digits
TableS5_Phys.Pair$Estimate<-round(TableS5_Phys.Pair$Estimate, 3)
TableS5_Phys.Pair$SE<-round(TableS5_Phys.Pair$SE, 3)
TableS5_Phys.Pair$t<-round(TableS5_Phys.Pair$t, 3)
TableS5_Phys.Pair$p<-round(TableS5_Phys.Pair$p, 3)
TableS5_Phys.Pair$p_Timepoint<-round(TableS5_Phys.Pair$p_Timepoint, 3)
TableS5_Phys.Pair$EffectSize<-round(TableS5_Phys.Pair$EffectSize, 3)
TableS5_Phys.Pair$EffectSize.SE<-round(TableS5_Phys.Pair$EffectSize.SE, 3)

#Integer
TableS5_Phys.Pair$DF<-round(TableS5_Phys.Pair$DF, 0)

##Write Out Table
write.csv(TableS5_Phys.Pair, "Tables/TableS5_Physiology_Pairwise_Results.csv", row.names=FALSE)

```
