==> 04_Performance.Rmd <==
---
title: "Coral Performance Following Reciprocal Transplant"
author: "Serena Hackerott and Carly Travers"
date: "1/3/2024"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  html_document:
    toc: yes
    df_print: paged
---

# Setup

```{r Setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


### Load Packages
```{r}
##Install Packages if Needed
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("cowplot")) install.packages("cowplot")
if (!require("lubridate")) install.packages("lubridate")
if (!require("corrplot")) install.packages("corrplot")
if (!require("lme4")) install.packages("lme4")
if (!require("vegan")) install.packages("vegan")
if (!require("DHARMa")) install.packages("DHARMa")
if (!require("effectsize")) install.packages("effectsize")
if (!require("emmeans")) install.packages("emmeans")
if (!require("lmerTest")) install.packages("lmerTest")
if (!require("tidyr")) install.packages("tidyr")
if (!require("dplyr")) install.packages("dplyr")
if (!require("plotrix")) install.packages("plotrix")
if (!require("Rmisc")) install.packages("Rmisc")
if (!require("ggpubr")) install.packages("ggpubr")

##Load Packages
library(ggplot2) #Required for ggplots
library(cowplot) #Required for plotting panel figures
library(lubridate) #Required for date and time formatting
library(corrplot) #Required for correlation plot
library(lme4) #Required for mixed effects modeling
library(vegan) #Required for multivariate analysis PERM
library(DHARMa) #Required to check residuals of mixed effects models
library(effectsize) #Required for eta_squared effect sizes
library(emmeans) #Required for pairwise comparisons 
library(lmerTest) #Required for p values with lme4 model summaries
library(tidyr) #Required for pivot_wider function
library(dplyr) #Required for group_by function
library(plotrix) #Required for Standard error function
library(Rmisc) #Required for SummarySE function
library(ggpubr) #Required for stat brackets

```


### Graphing Parameters
```{r}
#Note: Run "Graphing Parameters" section from 01_ExperimentalSetup.Rmd file
```


# Sample Data and Metadata

### Load and Organize Data
```{r}
##Growth
Length<-read.csv("Data/LengthFull.csv", header=TRUE)
Dates<-read.csv("Data/BonaireDates.csv", header=TRUE)

##Set date variables
Dates$Date<-as.Date(Dates$Date, format='%m/%d/%Y')

##Set factor variables
Length$Site<-factor(Length$Site, levels=c("KL", "SS"))
Length$Genotype<-factor(Length$Genotype, levels=c("AC8", "AC10", "AC12"))
Length$Orig<-factor(Length$Orig, levels=c("N", "T"))
Length$Origin<-factor(Length$Origin, levels=c("Native", "Transplant"))

##Add a Sample Set Variable
Length$Set<-paste(Length$Site, Length$Genotype, Length$Orig, sep=".")
Length$Set<-factor(Length$Set)

##Add Site.Orig variable
Length$Site.Orig<-paste(Length$Site, Length$Orig, sep=".")
Length$Site.Orig<-factor(Length$Site.Orig, levels=c("KL.N", "KL.T", "SS.N", "SS.T"))


##Bleaching Metrics
#Note: Physiological metrics calculated in 02_PhysiologyMetrics.R file
Thermal<-read.csv("Outputs/ThermData.csv", header=TRUE)

##Set factor variables
Thermal$TimeP<-factor(Thermal$TimeP, levels=c("IN", "TP1", "TP2", "TP3", "TP4"))
Thermal$Site<-factor(Thermal$Site, levels=c("KL", "SS"))
Thermal$Genotype<-factor(Thermal$Genotype, levels=c("AC8", "AC10", "AC12"))
Thermal$Orig<-factor(Thermal$Orig, levels=c("N", "T"))
Thermal$Origin<-factor(Thermal$Origin, levels=c("Native", "Transplant"))

```


# Growth

### Growth Rate TP1 to TP2

#### Calculate Days
```{r}
##Subset Dates for TP 1 and TP 2 Growth
Growth.Dates_TP1.2<-subset(Dates, Task=="Growth" & TimeP=="TP1" |
                       Task=="Growth" & TimeP=="TP2" )

##Convert to Wide format
Growth.Dates_TP1.2<-Growth.Dates_TP1.2 %>%  pivot_wider(names_from = TimeP, values_from = Date)

##Calculate Number of Days
#TP 1 to TP2
Growth.Dates_TP1.2$TP1.2_days<-time_length(Growth.Dates_TP1.2$TP2-Growth.Dates_TP1.2$TP1, unit="days")

```


#### Calculate Total Linear Extension
```{r}
##Merge Length Data with Growth Dates
#Merges by Site column
#Adds TP1.2_days column
Growth_TP1.2<-merge(Length, Growth.Dates_TP1.2[,c(-2)], all.x=TRUE)

##Calculate Linear Extension (cm)
Growth_TP1.2$Ext_cm<-Growth_TP1.2$TL.TP2_cm-Growth_TP1.2$TL.TP1_cm

##Calculate Growth_TP1.2 Rate (cm/day)
Growth_TP1.2$TLE_cm.day<-Growth_TP1.2$Ext_cm/Growth_TP1.2$TP1.2_days

```


#### Data QC
```{r}
##Remove Corals not Measured in T1 to T2
Growth_TP1.2<-Growth_TP1.2 %>% drop_na(TLE_cm.day)

##Sample Size per Set
Growth_TP1.2 %>% 
  dplyr::group_by(Set) %>%
  dplyr::summarize(SampleSize = length(Set))

##Check for Outliers
ggplot(Growth_TP1.2, aes(x=Set, y=TLE_cm.day)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

```

Potential outliers have been re-measured and therefore will be retained. n=10 per Site, Genotype, Origin group


### Compare Growth Rate
```{r}
#Check normality
hist(Growth_TP1.2$TLE_cm.day)
shapiro.test(Growth_TP1.2$TLE_cm.day)
#Not normal

hist(log(Growth_TP1.2$TLE_cm.day+1))
shapiro.test(log(Growth_TP1.2$TLE_cm.day+1))
##Still not normal

##Compare generalized linear mixed effects models
TLE_TP1.2.glmr.gaus<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=gaussian(link="identity"))
TLE_TP1.2.glmr.gaus.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=gaussian(link="log"))
TLE_TP1.2.glmr.gaus.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=gaussian(link="inverse"))

TLE_TP1.2.glmr.gam.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=Gamma(link="log"))
TLE_TP1.2.glmr.gam.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=Gamma(link="inverse"))

TLE_TP1.2.glmr.inv<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=inverse.gaussian(link="1/mu^2"))
TLE_TP1.2.glmr.inv.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=inverse.gaussian(link="log"))
TLE_TP1.2.glmr.inv.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP1.2, family=inverse.gaussian(link="inverse"))

AIC(TLE_TP1.2.glmr.gaus, TLE_TP1.2.glmr.gaus.l, TLE_TP1.2.glmr.gaus.i, TLE_TP1.2.glmr.gam.l, TLE_TP1.2.glmr.gam.i, TLE_TP1.2.glmr.inv, TLE_TP1.2.glmr.inv.l, TLE_TP1.2.glmr.inv.i)
```

Gamma distribution with log-link has the lowest AIC.


```{r}
##Check residuals
TLE_TP1.2.glmr.gam.l_res <- simulateResiduals(fittedModel = TLE_TP1.2.glmr.gam.l, plot = F)
plot(TLE_TP1.2.glmr.gam.l_res)

```


Compare residuals across models

```{r}
##Check residuals
TLE_TP1.2.glmr.gaus.l_res <- simulateResiduals(fittedModel = TLE_TP1.2.glmr.gaus.l, plot = F)
plot(TLE_TP1.2.glmr.gaus.l_res)

```

Compare with log+1 transformed lmer model

#### LMER Model
```{r}
##Model with log +1 transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
TLE_TP1.2.lme<-lmer(log(TLE_cm.day+1)~Origin*Site+(1|Genotype), data=Growth_TP1.2)

##Check residuals
TLE_TP1.2.lme_res <- simulateResiduals(fittedModel = TLE_TP1.2.lme, plot = F)
plot(TLE_TP1.2.lme_res)
```
Better residuals with the log+1 transformed lmer model.


```{r}
##Model results
summary(TLE_TP1.2.lme)

eta_squared(TLE_TP1.2.lme)

##Save model results
TLE_TP1.2.res<-data.frame(summary(TLE_TP1.2.lme)$coefficients[-1,])
TLE_TP1.2.res$Predictor<-c("Origin", "Site", "Origin x Site")
TLE_TP1.2.res$EtaSq<-c(eta_squared(TLE_TP1.2.lme)$Eta2)
TLE_TP1.2.res$Response<-rep("Growth", nrow(TLE_TP1.2.res))
TLE_TP1.2.res$Timepoint<-rep("TP1_2", nrow(TLE_TP1.2.res))

```

#### Pairwise
```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
TLE.pair_TP1.2<-emmeans(TLE_TP1.2.lme, pairwise~Origin | Site)
TLE.pair_TP1.2

#Calculate standardized effect sizes for pairwise comparisons
TLE.pair.es_TP1.2<-data.frame(eff_size(TLE.pair_TP1.2, sigma=sigma(TLE_TP1.2.lme), edf=df.residual(TLE_TP1.2.lme)))
TLE.pair.es_TP1.2
TLE.pair.es_TP1.2<-TLE.pair.es_TP1.2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
TLE.pair_TP1.2.res<-merge(data.frame(TLE.pair_TP1.2$contrasts), TLE.pair.es_TP1.2[,-c(1)])
TLE.pair_TP1.2.res$Response<-rep("Growth", nrow(TLE.pair_TP1.2.res))
TLE.pair_TP1.2.res$TimeP<-rep("TP1_2", nrow(TLE.pair_TP1.2.res))

##Add Significance levels
TLE.pair_TP1.2.res$Sig<-ifelse(TLE.pair_TP1.2.res$p.value<0.001, "***",
                              ifelse(TLE.pair_TP1.2.res$p.value<0.01, "**",
                                     ifelse(TLE.pair_TP1.2.res$p.value<0.05, "*",
                                            ifelse(TLE.pair_TP1.2.res$p.value<0.1, "-", NA))))
TLE.pair_TP1.2.res
```


#### Plot Growth TP 1 to 2
```{r Plot Growth Rate TP1 to 2}
##Summary statistics by Site and Origin
TP1.2_Growth.sum<-summarySE(Growth_TP1.2, measurevar="TLE_cm.day", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Growth Rate across Treatments
TP1.2_Growth.plot<-ggplot(TP1.2_Growth.sum, aes(x=Site, y=TLE_cm.day, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TLE_cm.day-se, ymax=TLE_cm.day+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y=expression(paste('Growth Rate (cm day'^-1*")")), colour="Origin")+
  ylim(0, 0.25)+
  annotate("text", x=2, y=0.24, label="*", size=sig.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=0.01, label.size=levels.sz, label="**", inherit.aes = FALSE); TP1.2_Growth.plot
```


### Percent Difference in Growth Rates
```{r}
##Summary statistics by Genotype, Origin, and Site
TP1.2_Growth.sum<-summarySE(Growth_TP1.2, measurevar="TLE_cm.day", groupvars=c("Genotype", "Origin", "Site"), na.rm=TRUE)

##Calculate Percent Difference between Native and Transplant by Genotype
TP1.2_Growth.dif<-TP1.2_Growth.sum[which(TP1.2_Growth.sum$Origin=="Native"), c(1,3,5)]
TP1.2_Growth.dif<-TP1.2_Growth.dif %>% dplyr::rename(TLE_N = TLE_cm.day)
TP1.2_Growth.dif$TLE_T<-TP1.2_Growth.sum$TLE_cm.day[which(TP1.2_Growth.sum$Origin=="Transplant")]

##SS
TP1.2_Growth.dif_SS<-subset(TP1.2_Growth.dif, Site=="SS")
TP1.2_Growth.dif_SS$Perc.Inc<-((TP1.2_Growth.dif_SS$TLE_N-TP1.2_Growth.dif_SS$TLE_T)/TP1.2_Growth.dif_SS$TLE_T)*100
TP1.2_Growth.dif_SS

##Mean and Standard Error
mean(TP1.2_Growth.dif_SS$Perc.Inc)
std.error(TP1.2_Growth.dif_SS$Perc.Inc)

##KL
TP1.2_Growth.dif_KL<-subset(TP1.2_Growth.dif, Site=="KL")
TP1.2_Growth.dif_KL$Perc.Inc<-((TP1.2_Growth.dif_KL$TLE_N-TP1.2_Growth.dif_KL$TLE_T)/TP1.2_Growth.dif_KL$TLE_T)*100
TP1.2_Growth.dif_KL

##Mean and Standard Error
mean(TP1.2_Growth.dif_KL$Perc.Inc)
std.error(TP1.2_Growth.dif_KL$Perc.Inc)


```


### Growth Rate TP3 to TP4

#### Calculate Days
```{r}
##Subset Dates for TP 3 and TP 4 Growth
Growth.Dates_TP3.4<-subset(Dates, Task=="Growth" & TimeP=="TP3" |
                       Task=="Growth" & TimeP=="TP4" )

##Convert to Wide format
Growth.Dates_TP3.4<-Growth.Dates_TP3.4 %>%  pivot_wider(names_from = TimeP, values_from = Date)

##Calculate Number of Days
#TP 3 to TP4
Growth.Dates_TP3.4$TP3.4_days<-time_length(Growth.Dates_TP3.4$TP4-Growth.Dates_TP3.4$TP3, unit="days")

```


#### Calculate Total Linear Extension
```{r}
##Merge Length Data with Growth Dates
#Merges by Site column
#Adds TP3.4_days column
Growth_TP3.4<-merge(Length, Growth.Dates_TP3.4[,c(-2)], all.x=TRUE)

##Calculate Linear Extension (cm)
Growth_TP3.4$Ext_cm<-Growth_TP3.4$TL.TP4_cm-Growth_TP3.4$TL.TP3_cm

##Calculate Growth_TP3.4 Rate (cm/day)
Growth_TP3.4$TLE_cm.day<-Growth_TP3.4$Ext_cm/Growth_TP3.4$TP3.4_days

```


#### Data QC
```{r}
##Remove Corals not Measured in T1 to T2
Growth_TP3.4<-Growth_TP3.4 %>% drop_na(TLE_cm.day)

##Sample Size per Set
Growth_TP3.4 %>% 
  dplyr::group_by(Set) %>%
  dplyr::summarize(SampleSize = length(Set))

##Check for Outliers
ggplot(Growth_TP3.4, aes(x=Set, y=TLE_cm.day)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

```


### Compare Growth Rate
```{r}
#Check normality
hist(Growth_TP3.4$TLE_cm.day)
shapiro.test(Growth_TP3.4$TLE_cm.day)
#Not normal

hist(log(Growth_TP3.4$TLE_cm.day+1))
shapiro.test(log(Growth_TP3.4$TLE_cm.day+1))
##Still not normal

##Compare generalized linear mixed effects models
TLE_TP3.4.glmr.gaus<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=gaussian(link="identity"))
TLE_TP3.4.glmr.gaus.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=gaussian(link="log"))
TLE_TP3.4.glmr.gaus.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=gaussian(link="inverse"))

TLE_TP3.4.glmr.gam.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=Gamma(link="log"))
TLE_TP3.4.glmr.gam.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=Gamma(link="inverse"))

TLE_TP3.4.glmr.inv<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=inverse.gaussian(link="1/mu^2"))
TLE_TP3.4.glmr.inv.l<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=inverse.gaussian(link="log"))
TLE_TP3.4.glmr.inv.i<-glmer(TLE_cm.day~Origin*Site+(1|Genotype), data=Growth_TP3.4, family=inverse.gaussian(link="inverse"))

AIC(TLE_TP3.4.glmr.gaus, TLE_TP3.4.glmr.gaus.l, TLE_TP3.4.glmr.gaus.i, TLE_TP3.4.glmr.gam.l, TLE_TP3.4.glmr.gam.i, TLE_TP3.4.glmr.inv, TLE_TP3.4.glmr.inv.l, TLE_TP3.4.glmr.inv.i)
```

Gamma distribution with log-link has the lowest AIC.


```{r}
##Check residuals
TLE_TP3.4.glmr.gam.l_res <- simulateResiduals(fittedModel = TLE_TP3.4.glmr.gam.l, plot = F)
plot(TLE_TP3.4.glmr.gam.l_res)

```


Compare residuals across models

```{r}
##Check residuals
TLE_TP3.4.glmr.gaus.l_res <- simulateResiduals(fittedModel = TLE_TP3.4.glmr.gaus.l, plot = F)
plot(TLE_TP3.4.glmr.gaus.l_res)

```

Compare with log+1 transformed lmer model

#### LMER Model
```{r}
##Model with log +1 transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
TLE_TP3.4.lme<-lmer(log(TLE_cm.day+1)~Origin*Site+(1|Genotype), data=Growth_TP3.4)

##Check residuals
TLE_TP3.4.lme_res <- simulateResiduals(fittedModel = TLE_TP3.4.lme, plot = F)
plot(TLE_TP3.4.lme_res)
```
Better residuals with the log+1 transformed lmer model.


```{r}
##Model results
summary(TLE_TP3.4.lme)

eta_squared(TLE_TP3.4.lme)

##Save model results
TLE_TP3.4.res<-data.frame(summary(TLE_TP3.4.lme)$coefficients[-1,])
TLE_TP3.4.res$Predictor<-c("Origin", "Site", "Origin x Site")
TLE_TP3.4.res$EtaSq<-c(eta_squared(TLE_TP3.4.lme)$Eta2)
TLE_TP3.4.res$Response<-rep("Growth", nrow(TLE_TP3.4.res))
TLE_TP3.4.res$Timepoint<-rep("TP3_4", nrow(TLE_TP3.4.res))

```

#### Pairwise 
```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
TLE.pair_TP3.4<-emmeans(TLE_TP3.4.lme, pairwise~Origin | Site)
TLE.pair_TP3.4

#Calculate standardized effect sizes for pairwise comparisons
TLE.pair.es_TP3.4<-data.frame(eff_size(TLE.pair_TP3.4, sigma=sigma(TLE_TP3.4.lme), edf=df.residual(TLE_TP3.4.lme)))
TLE.pair.es_TP3.4
TLE.pair.es_TP3.4<-TLE.pair.es_TP3.4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
TLE.pair_TP3.4.res<-merge(data.frame(TLE.pair_TP3.4$contrasts), TLE.pair.es_TP3.4[,-c(1)])
TLE.pair_TP3.4.res$Response<-rep("Growth", nrow(TLE.pair_TP3.4.res))
TLE.pair_TP3.4.res$TimeP<-rep("TP3_4", nrow(TLE.pair_TP3.4.res))

##Add Significance levels
TLE.pair_TP3.4.res$Sig<-ifelse(TLE.pair_TP3.4.res$p.value<0.001, "***",
                              ifelse(TLE.pair_TP3.4.res$p.value<0.01, "**",
                                     ifelse(TLE.pair_TP3.4.res$p.value<0.05, "*",
                                            ifelse(TLE.pair_TP3.4.res$p.value<0.1, "-", NA))))
TLE.pair_TP3.4.res
```


#### Plot Growth TP 3 to 4
```{r Plot Growth Rate TP3 to 4}
##Summary statistics by Site and Origin
TP3.4_Growth.sum<-summarySE(Growth_TP3.4, measurevar="TLE_cm.day", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Growth Rate across Treatments
TP3.4_Growth.plot<-ggplot(TP3.4_Growth.sum, aes(x=Site, y=TLE_cm.day, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=TLE_cm.day-se, ymax=TLE_cm.day+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y=expression(paste('Growth Rate (cm day'^-1*")")), colour="Origin")+
  ylim(0, 0.25)+
  annotate("text", x=1, y=0.22, label="*", size=sig.sz, fontface="bold")+
  annotate("text", x=2, y=0.18, label="**", size=sig.sz, fontface="bold")+
  geom_bracket(xmin=1, xmax=2, y.position=0.01, label.size=levels.sz, label="***", inherit.aes = FALSE); TP3.4_Growth.plot
```


### Percent Difference in Growth Rates
```{r}
##Summary statistics by Genotype, Origin, and Site
TP3.4_Growth.sum<-summarySE(Growth_TP3.4, measurevar="TLE_cm.day", groupvars=c("Genotype", "Origin", "Site"), na.rm=TRUE)

##Calculate Percent Difference between Native and Transplant by Genotype
TP3.4_Growth.dif<-TP3.4_Growth.sum[which(TP3.4_Growth.sum$Origin=="Native"), c(1,3,5)]
TP3.4_Growth.dif<-TP3.4_Growth.dif %>% dplyr::rename(TLE_N = TLE_cm.day)
TP3.4_Growth.dif$TLE_T<-TP3.4_Growth.sum$TLE_cm.day[which(TP3.4_Growth.sum$Origin=="Transplant")]

##SS
TP3.4_Growth.dif_SS<-subset(TP3.4_Growth.dif, Site=="SS")
TP3.4_Growth.dif_SS$Perc.Inc<-((TP3.4_Growth.dif_SS$TLE_T-TP3.4_Growth.dif_SS$TLE_N)/TP3.4_Growth.dif_SS$TLE_N)*100
TP3.4_Growth.dif_SS

##Mean and Standard Error
mean(TP3.4_Growth.dif_SS$Perc.Inc)
std.error(TP3.4_Growth.dif_SS$Perc.Inc)

##KL
TP3.4_Growth.dif_KL<-subset(TP3.4_Growth.dif, Site=="KL")
TP3.4_Growth.dif_KL$Perc.Inc<-((TP3.4_Growth.dif_KL$TLE_N-TP3.4_Growth.dif_KL$TLE_T)/TP3.4_Growth.dif_KL$TLE_T)*100
TP3.4_Growth.dif_KL

##Mean and Standard Error
mean(TP3.4_Growth.dif_KL$Perc.Inc)
std.error(TP3.4_Growth.dif_KL$Perc.Inc)

```


# Write Out Growth Data
```{r}
##Growth Data dataframe
GrowthData<-Growth_TP1.2

##Specify Timepoints
GrowthData<-GrowthData %>% dplyr::rename(TP1.2_Ext_cm = Ext_cm, TP1.2_TLE_cm.day = TLE_cm.day)

##Merge with Timepoints 3 to 4
GrowthData<-merge(GrowthData, Growth_TP3.4, all=TRUE)

##Specify Timepoints
GrowthData<-GrowthData %>% dplyr::rename(TP3.4_Ext_cm = Ext_cm, TP3.4_TLE_cm.day = TLE_cm.day)

##Growth Data
write.csv(GrowthData, "Outputs/GrowthData.csv", row.names=FALSE)
```


# Thermal Tolerance

### Percent Retention

#### Subset Thermal Assay Data by Treatment
```{r}
##Control
Therm_C<-subset(Thermal, Treat=="C")

##Heated
Therm_H<-subset(Thermal, Treat=="H")
```


#### Calculate Percent Retention
Calculating retention as proportion remaining relative to corresponding control levels (0-1) for each site and genotype at each timepoint.
```{r}
##Calculate averages for Control Treatment for each Site, Genotype, and Timepoint 
Therm_C<-na.omit(Therm_C)
names(Therm_C)
Therm_C.a<-aggregate(Therm_C[,c(13:15)], list(Therm_C$Site, Therm_C$Genotype, Therm_C$TimeP, Therm_C$Origin), mean, na.action = na.omit)
names(Therm_C.a)[1:4]<-c("Site", "Genotype", "TimeP", "Origin")
names(Therm_C.a)[5:7]<-paste(names(Therm_C.a)[5:7], "C", sep="_")

##Merge Control Averages with Heated Samples
#Merges by Site, Genotype, and Timepoint
Therm_H<-merge(Therm_H, Therm_C.a, all.x=TRUE )

##Calculate Proportion Retained for each Bleaching Metric relative to Control
Therm_H$Chl.prop<-round((Therm_H$Chl_ug.cm2/Therm_H$Chl_ug.cm2_C), 4)
Therm_H$Sym.prop<-round((Therm_H$Sym10.6_cm2/Therm_H$Sym10.6_cm2_C), 4)
Therm_H$PAM.prop<-round((Therm_H$Fv_Fm/Therm_H$Fv_Fm_C), 4)

##Set values >1 to 1
Therm_H$Chl.prop[which(Therm_H$Chl.prop>1)]<-1.0000
Therm_H$Sym.prop[which(Therm_H$Sym.prop>1)]<-1.0000
Therm_H$PAM.prop[which(Therm_H$PAM.prop>1)]<-1.0000

```


# Initial Heat Assay 

```{r}
##Subset Initial Timepoint
Thermal_IN<-subset(Thermal, TimeP=="IN")
Thermal_IN$Site.Treat<-paste(Thermal_IN$Site, Thermal_IN$Treat, sep=".")
Thermal_IN$Site.Treat<-factor(Thermal_IN$Site.Treat, levels=c("KL.C", "KL.H", "SS.C", "SS.H"))


Therm_IN<-subset(Therm_H, TimeP=="IN")
```


### Heated vs Control


#### Symbionts
```{r}
#Check normality
hist(Thermal_IN$Sym10.6_cm2)
shapiro.test(Thermal_IN$Sym10.6_cm2)
#Not normal

hist(log(Thermal_IN$Sym10.6_cm2+1))
shapiro.test(log(Thermal_IN$Sym10.6_cm2+1))
#Normal

##Model
#Function of Site and Treatment, with Genotype as a Random effect
Sym.lme_IN<-lmer(log(Sym10.6_cm2+1)~Site*Treatment+(1|Genotype), data=Thermal_IN)

##Check residuals
Sym.lme_res_IN <- simulateResiduals(fittedModel = Sym.lme_IN, plot = F)
plot(Sym.lme_res_IN)

##Model results
summary(Sym.lme_IN)

eta_squared(Sym.lme_IN)

emmeans(Sym.lme_IN, pairwise ~ Site | Treatment)

emmeans(Sym.lme_IN, pairwise ~ Treatment | Site)

##Save model results
Sym_IN.res<-data.frame(summary(Sym.lme_IN)$coefficients[-1,])

##Add pairwise
names(Sym_IN.res)<-names(data.frame(emmeans(Sym.lme_IN, pairwise ~ Site | Treatment)$contrasts))[-c(1:2)]
Sym_IN.res<-rbind(Sym_IN.res,
data.frame(rbind(emmeans(Sym.lme_IN, pairwise ~ Site | Treatment)$contrasts[1], emmeans(Sym.lme_IN, pairwise ~ Site | Treatment)$contrasts[2], emmeans(Sym.lme_IN, pairwise ~ Treatment | Site)$contrasts[1], emmeans(Sym.lme_IN, pairwise ~ Treatment | Site)$contrasts[2]))[,-c(1:3)])

##Metadata
Sym_IN.res$Predictor<-c("Site", "Treatment", "Site x Treatment", "Control Site", "Heated Site", "KL Treatment", "SS Treatment")
Sym_IN.res$Response<-rep("Symbionts", nrow(Sym_IN.res))
Sym_IN.res$EtaSq<-c(eta_squared(Sym.lme_IN)$Eta2, rep(NA, 4))

```


```{r Plot Symbionts Initial}
##Summary statistics by Site
IN_Sym.sum<-summarySE(Thermal_IN, measurevar="Sym10.6_cm2", groupvars=c("Site", "Treatment", "Site.Treat"), na.rm=TRUE)

##Plot Average Symbionts across Treatments
IN_Sym.plot<-ggplot(IN_Sym.sum, aes(x=Site.Treat, y=Sym10.6_cm2, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=Sym10.6_cm2-se, ymax=Sym10.6_cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(aes(shape=Treatment), size=point.sz, position=position_dodge(width=0.5))+
  scale_shape_manual(values=c(1,16))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Treatment", y=expression(paste('Symbiont Density ('*10^6,'cells cm'^-2*")")), colour="Site")+
  ylim(0, 1)+
  annotate("text", x=.6, y=1, hjust=0, label="Site **", size=sig.sz, fontface="bold")+
  annotate("text", x=.6, y=.94, hjust=0, label="Treatment **", size=sig.sz, fontface="bold"); IN_Sym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Thermal_IN$Chl_ug.cm2)
shapiro.test(Thermal_IN$Chl_ug.cm2)
#Not normal

hist(log(Thermal_IN$Chl_ug.cm2+1))
shapiro.test(log(Thermal_IN$Chl_ug.cm2+1))
#Still not normal but less skewed

##Model
#Function of Site and Treatment, with Genotype as a Random effect
Chl.lme_IN<-lmer(log(Chl_ug.cm2+1)~Site*Treatment+(1|Genotype), data=Thermal_IN)

##Check residuals
Chl.lme_res_IN <- simulateResiduals(fittedModel = Chl.lme_IN, plot = F)
plot(Chl.lme_res_IN)

##Model results
summary(Chl.lme_IN)

eta_squared(Chl.lme_IN)

emmeans(Chl.lme_IN, pairwise ~ Site | Treatment)

emmeans(Chl.lme_IN, pairwise ~ Treatment | Site)


##Save model results
Chl_IN.res<-data.frame(summary(Chl.lme_IN)$coefficients[-1,])

##Add pairwise
names(Chl_IN.res)<-names(data.frame(emmeans(Chl.lme_IN, pairwise ~ Site | Treatment)$contrasts))[-c(1:2)]
Chl_IN.res<-rbind(Chl_IN.res,
data.frame(rbind(emmeans(Chl.lme_IN, pairwise ~ Site | Treatment)$contrasts[1], emmeans(Chl.lme_IN, pairwise ~ Site | Treatment)$contrasts[2], emmeans(Chl.lme_IN, pairwise ~ Treatment | Site)$contrasts[1], emmeans(Chl.lme_IN, pairwise ~ Treatment | Site)$contrasts[2]))[,-c(1:3)])

##Metadata
Chl_IN.res$Predictor<-c("Site", "Treatment", "Site x Treatment", "Control Site", "Heated Site", "KL Treatment", "SS Treatment")
Chl_IN.res$Response<-rep("Chlorophyll", nrow(Chl_IN.res))
Chl_IN.res$EtaSq<-c(eta_squared(Chl.lme_IN)$Eta2, rep(NA, 4))

```


```{r Plot Chlorophyll Initial}
##Summary statistics by Site
IN_Chl.sum<-summarySE(Thermal_IN, measurevar="Chl_ug.cm2", groupvars=c("Site", "Treatment", "Site.Treat"), na.rm=TRUE)

##Plot Average Chlorophyll across Treatments
IN_Chl.plot<-ggplot(IN_Chl.sum, aes(x=Site.Treat, y=Chl_ug.cm2, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=Chl_ug.cm2-se, ymax=Chl_ug.cm2+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(aes(shape=Treatment), size=point.sz, position=position_dodge(width=0.5))+
  scale_shape_manual(values=c(1,16))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Treatment", y=expression(paste('Total Chlorophyll (\u03BCg cm'^-2*")")), colour="Site")+
  ylim(0, 1.25)+
  annotate("text", x=.6, y=1.2, hjust=0, label="Site ***", size=sig.sz, fontface="bold")+
  annotate("text", x=.6, y=1.12, hjust=0, label="Treatment ***", size=sig.sz, fontface="bold")+
  annotate("text", x=.6, y=1.04, hjust=0, label="Site x Treatment **", size=sig.sz, fontface="bold"); IN_Chl.plot
```


#### FvFm
```{r}
#Check normality
hist(Thermal_IN$Fv_Fm)
shapiro.test(Thermal_IN$Fv_Fm)
#Not normal

hist(Thermal_IN$Fv_Fm^2)
shapiro.test(Thermal_IN$Fv_Fm^2)
#Still not normal but less skewed

##Model
#Function of Site and Treatment, with Genotype as a Random effect
PAM.lme_IN<-lmer(Fv_Fm^2~Site*Treatment+(1|Genotype), data=Thermal_IN)

##Check residuals
PAM.lme_res_IN <- simulateResiduals(fittedModel = PAM.lme_IN, plot = F)
plot(PAM.lme_res_IN)

##Model results
summary(PAM.lme_IN)

eta_squared(PAM.lme_IN)

emmeans(PAM.lme_IN, pairwise ~ Site | Treatment)

emmeans(PAM.lme_IN, pairwise ~ Treatment | Site)


##Save model results
PAM_IN.res<-data.frame(summary(PAM.lme_IN)$coefficients[-1,])

##Add pairwise
names(PAM_IN.res)<-names(data.frame(emmeans(PAM.lme_IN, pairwise ~ Site | Treatment)$contrasts))[-c(1:2)]
PAM_IN.res<-rbind(PAM_IN.res,
data.frame(rbind(emmeans(PAM.lme_IN, pairwise ~ Site | Treatment)$contrasts[1], emmeans(PAM.lme_IN, pairwise ~ Site | Treatment)$contrasts[2], emmeans(PAM.lme_IN, pairwise ~ Treatment | Site)$contrasts[1], emmeans(PAM.lme_IN, pairwise ~ Treatment | Site)$contrasts[2]))[,-c(1:3)])

##Metadata
PAM_IN.res$Predictor<-c("Site", "Treatment", "Site x Treatment", "Control Site", "Heated Site", "KL Treatment", "SS Treatment")
PAM_IN.res$Response<-rep("FvFm", nrow(PAM_IN.res))
PAM_IN.res$EtaSq<-c(eta_squared(PAM.lme_IN)$Eta2, rep(NA, 4))

```
Not normal, but model residuals are OK.


```{r Plot FvFm Initial}
##Summary statistics by Site
IN_PAM.sum<-summarySE(Thermal_IN, measurevar="Fv_Fm", groupvars=c("Site", "Treatment", "Site.Treat"), na.rm=TRUE)

##Plot Average FvFm across Treatments
IN_PAM.plot<-ggplot(IN_PAM.sum, aes(x=Site.Treat, y=Fv_Fm, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=Fv_Fm-se, ymax=Fv_Fm+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(aes(shape=Treatment), size=point.sz, position=position_dodge(width=0.5))+
  scale_shape_manual(values=c(1,16))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Treatment", y="Photochemical Efficiency (Fv/Fm)", colour="Site")+
  ylim(0.5, 0.655)+
  annotate("text", x=.6, y=.655, hjust=0, label="Site *", size=sig.sz, fontface="bold")+
  annotate("text", x=.6, y=.645, hjust=0, label="Treatment ***", size=sig.sz, fontface="bold"); IN_PAM.plot
```


### Retention by Site

#### Symbionts
```{r}
#Check normality
hist(Therm_IN$Sym.prop)
shapiro.test(Therm_IN$Sym.prop)
#Normal

##Model
#Function of Site, with Genotype as a Random effect
Tol_Sym.lme_IN<-lmer(Sym.prop~Site+(1|Genotype), data=Therm_IN)

##Check residuals
Tol_Sym.lme_res_IN <- simulateResiduals(fittedModel = Tol_Sym.lme_IN, plot = F)
plot(Tol_Sym.lme_res_IN)

##Model results
summary(Tol_Sym.lme_IN)

eta_squared(Tol_Sym.lme_IN)

##Save model results
Tol_Sym_IN.res<-data.frame(summary(Tol_Sym.lme_IN)$coefficients)
names(Tol_Sym_IN.res)<-names(Sym_IN.res)[1:5]
Tol_Sym_IN.res$Predictor<-rep("Site", nrow(Tol_Sym_IN.res))
Tol_Sym_IN.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym_IN.res))
Tol_Sym_IN.res$EtaSq<-c(eta_squared(Tol_Sym.lme_IN)$Eta2)

##Combine results
Sym_IN.res<-rbind(Sym_IN.res, Tol_Sym_IN.res[2,])
```


```{r Plot Symbiont Retention Initial}
##Summary statistics by Site
IN_TolSym.sum<-summarySE(Therm_IN, measurevar="Sym.prop", groupvars=c("Site"), na.rm=TRUE)

##Plot Average Symbiont Retention across Treatments
IN_TolSym.plot<-ggplot(IN_TolSym.sum, aes(x=Site, y=Sym.prop, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=Sym.prop-se, ymax=Sym.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site", y="Symbiont Retention", colour="Site")+
  ylim(0, 1)+
  annotate("text", x=1.5, y=0.75, label="-", size=sig.sz, fontface="bold"); IN_TolSym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Therm_IN$Chl.prop)
shapiro.test(Therm_IN$Chl.prop)
#Normal

##Model
#Function of Site, with Genotype as a Random effect
Tol_Chl.lme_IN<-lmer(Chl.prop~Site+(1|Genotype), data=Therm_IN)

##Check residuals
Tol_Chl.lme_res_IN <- simulateResiduals(fittedModel = Tol_Chl.lme_IN, plot = F)
plot(Tol_Chl.lme_res_IN)

##Model results
summary(Tol_Chl.lme_IN)

eta_squared(Tol_Chl.lme_IN)

##Save model results
Tol_Chl_IN.res<-data.frame(summary(Tol_Chl.lme_IN)$coefficients)
names(Tol_Chl_IN.res)<-names(Chl_IN.res)[1:5]
Tol_Chl_IN.res$Predictor<-rep("Site", nrow(Tol_Chl_IN.res))
Tol_Chl_IN.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl_IN.res))
Tol_Chl_IN.res$EtaSq<-c(eta_squared(Tol_Chl.lme_IN)$Eta2)

##Combine results
Chl_IN.res<-rbind(Chl_IN.res, Tol_Chl_IN.res[2,])
```


```{r Plot Chlorophyll Retention Initial}
##Summary statistics by Site and Origin
IN_TolChl.sum<-summarySE(Therm_IN, measurevar="Chl.prop", groupvars=c("Site"), na.rm=TRUE)

##Plot Average Chlorophyll Retention across Treatments
IN_TolChl.plot<-ggplot(IN_TolChl.sum, aes(x=Site, y=Chl.prop, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=Chl.prop-se, ymax=Chl.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site", y="Chlorophyll Retention", colour="Site")+
  ylim(0, 0.35); IN_TolChl.plot
```


#### FvFm
```{r}
#Check normality
hist(Therm_IN$PAM.prop)
shapiro.test(Therm_IN$PAM.prop)
#Not normal

hist(Therm_IN$PAM.prop^2)
shapiro.test(Therm_IN$PAM.prop^2)
##Still not normal

##Model
#Function of Site, with Genotype as a Random effect
Tol_PAM.lme_IN<-lmer(PAM.prop~Site+(1|Genotype), data=Therm_IN)

##Check residuals
Tol_PAM.lme_res_IN <- simulateResiduals(fittedModel = Tol_PAM.lme_IN, plot = F)
plot(Tol_PAM.lme_res_IN)

##Model results
summary(Tol_PAM.lme_IN)

eta_squared(Tol_PAM.lme_IN)

##Save model results
Tol_PAM_IN.res<-data.frame(summary(Tol_PAM.lme_IN)$coefficients)
names(Tol_PAM_IN.res)<-names(PAM_IN.res)[1:5]
Tol_PAM_IN.res$Predictor<-rep("Site", nrow(Tol_PAM_IN.res))
Tol_PAM_IN.res$Response<-rep("FvFm Retention", nrow(Tol_PAM_IN.res))
Tol_PAM_IN.res$EtaSq<-c(eta_squared(Tol_PAM.lme_IN)$Eta2)

##Combine results
PAM_IN.res<-rbind(PAM_IN.res, Tol_PAM_IN.res[2,])
```
Not normal, but model residuals are OK.


```{r Plot FvFm Retention Initial}
##Summary statistics by Site and Origin
IN_TolPAM.sum<-summarySE(Therm_IN, measurevar="PAM.prop", groupvars=c("Site"), na.rm=TRUE)

##Plot Average FvFm Retention across Treatments
IN_TolPAM.plot<-ggplot(IN_TolPAM.sum, aes(x=Site, y=PAM.prop, colour=Site)) + 
  scale_colour_manual(values=Site.colors.o)+
  geom_errorbar(aes(ymin=PAM.prop-se, ymax=PAM.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site", y="Photochemical Efficiency Retention", colour="Site")+
  ylim(0.6, 1)+
  annotate("text", x=1.5, y=1, label="*", size=sig.sz, fontface="bold"); IN_TolPAM.plot
```


# Tolerance after Transplant

### TP1
```{r}
##Subset Timepoint 1
Therm_TP1<-subset(Therm_H, TimeP=="TP1")
```


#### Symbionts
```{r}
#Check normality
hist(Therm_TP1$Sym.prop)
shapiro.test(Therm_TP1$Sym.prop)
#Not normal

hist(log(Therm_TP1$Sym.prop+1))
shapiro.test(log(Therm_TP1$Sym.prop+1))
##Still not normal 

##Model with no transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Sym.lme_TP1<-lmer(Sym.prop~Origin*Site+(1|Genotype), data=Therm_TP1)

##Check residuals
Tol_Sym.lme_res_TP1 <- simulateResiduals(fittedModel = Tol_Sym.lme_TP1, plot = F)
plot(Tol_Sym.lme_res_TP1)

##Model results
summary(Tol_Sym.lme_TP1)

eta_squared(Tol_Sym.lme_TP1)

##Save model results
Tol_Sym_TP1.res<-data.frame(summary(Tol_Sym.lme_TP1)$coefficients[-1,])
Tol_Sym_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Sym_TP1.res$EtaSq<-c(eta_squared(Tol_Sym.lme_TP1)$Eta2)
Tol_Sym_TP1.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym_TP1.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Sym.pair_TP1<-emmeans(Tol_Sym.lme_TP1, pairwise~Origin | Site)
Tol_Sym.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Tol_Sym.pair.es_TP1<-data.frame(eff_size(Tol_Sym.pair_TP1, sigma=sigma(Tol_Sym.lme_TP1), edf=df.residual(Tol_Sym.lme_TP1)))
Tol_Sym.pair.es_TP1
Tol_Sym.pair.es_TP1<-Tol_Sym.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Sym.pair_TP1.res<-merge(data.frame(Tol_Sym.pair_TP1$contrasts), Tol_Sym.pair.es_TP1[,-c(1)])
Tol_Sym.pair_TP1.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym.pair_TP1.res))

##Add Significance levels
Tol_Sym.pair_TP1.res$Sig<-ifelse(Tol_Sym.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Tol_Sym.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Tol_Sym.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Tol_Sym.pair_TP1.res$p.value<0.1, "-", NA))))
Tol_Sym.pair_TP1.res
```


```{r Plot Symbiont Retention TP1}
##Summary statistics by Site and Origin
TP1_TolSym.sum<-summarySE(Therm_TP1, measurevar="Sym.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Symbiont Retention across Treatments
TP1_TolSym.plot<-ggplot(TP1_TolSym.sum, aes(x=Site, y=Sym.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Sym.prop-se, ymax=Sym.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Symbiont Retention", colour="Origin")+
  ylim(0, 1)+
  annotate("text", x=1, y=0.65, label="-", size=sig.sz, fontface="bold"); TP1_TolSym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Therm_TP1$Chl.prop)
shapiro.test(Therm_TP1$Chl.prop)
#Not normal

hist(log(Therm_TP1$Chl.prop+1))
shapiro.test(log(Therm_TP1$Chl.prop+1))
##Still not normal but less skewed

##Model with no transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Chl.lme_TP1<-lmer(Chl.prop~Origin*Site+(1|Genotype), data=Therm_TP1)

##Check residuals
Tol_Chl.lme_res_TP1 <- simulateResiduals(fittedModel = Tol_Chl.lme_TP1, plot = F)
plot(Tol_Chl.lme_res_TP1)

##Model results
summary(Tol_Chl.lme_TP1)

eta_squared(Tol_Chl.lme_TP1)

##Save model results
Tol_Chl_TP1.res<-data.frame(summary(Tol_Chl.lme_TP1)$coefficients[-1,])
Tol_Chl_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Chl_TP1.res$EtaSq<-c(eta_squared(Tol_Chl.lme_TP1)$Eta2)
Tol_Chl_TP1.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl_TP1.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Chl.pair_TP1<-emmeans(Tol_Chl.lme_TP1, pairwise~Origin | Site)
Tol_Chl.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Tol_Chl.pair.es_TP1<-data.frame(eff_size(Tol_Chl.pair_TP1, sigma=sigma(Tol_Chl.lme_TP1), edf=df.residual(Tol_Chl.lme_TP1)))
Tol_Chl.pair.es_TP1
Tol_Chl.pair.es_TP1<-Tol_Chl.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Chl.pair_TP1.res<-merge(data.frame(Tol_Chl.pair_TP1$contrasts), Tol_Chl.pair.es_TP1[,-c(1)])
Tol_Chl.pair_TP1.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl.pair_TP1.res))

##Add Significance levels
Tol_Chl.pair_TP1.res$Sig<-ifelse(Tol_Chl.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Tol_Chl.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Tol_Chl.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Tol_Chl.pair_TP1.res$p.value<0.1, "-", NA))))
Tol_Chl.pair_TP1.res
```


```{r Plot Chlorophyll Retention TP1}
##Summary statistics by Site and Origin
TP1_TolChl.sum<-summarySE(Therm_TP1, measurevar="Chl.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Chlorophyll Retention across Treatments
TP1_TolChl.plot<-ggplot(TP1_TolChl.sum, aes(x=Site, y=Chl.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl.prop-se, ymax=Chl.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Chlorophyll Retention", colour="Origin")+
  ylim(0, 0.35); TP1_TolChl.plot
```


#### FvFm
```{r}
#Check normality
hist(Therm_TP1$PAM.prop)
shapiro.test(Therm_TP1$PAM.prop)
#Not normal

hist(Therm_TP1$PAM.prop^2)
shapiro.test(Therm_TP1$PAM.prop^2)
##Still not normal

##Model with no transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_PAM.lme_TP1<-lmer(PAM.prop~Origin*Site+(1|Genotype), data=Therm_TP1)

##Check residuals
Tol_PAM.lme_res_TP1 <- simulateResiduals(fittedModel = Tol_PAM.lme_TP1, plot = F)
plot(Tol_PAM.lme_res_TP1)

##Model results
summary(Tol_PAM.lme_TP1)

eta_squared(Tol_PAM.lme_TP1)

##Save model results
Tol_PAM_TP1.res<-data.frame(summary(Tol_PAM.lme_TP1)$coefficients[-1,])
Tol_PAM_TP1.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_PAM_TP1.res$EtaSq<-c(eta_squared(Tol_PAM.lme_TP1)$Eta2)
Tol_PAM_TP1.res$Response<-rep("FvFm Retention", nrow(Tol_PAM_TP1.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_PAM.pair_TP1<-emmeans(Tol_PAM.lme_TP1, pairwise~Origin | Site)
Tol_PAM.pair_TP1

#Calculate standardized effect sizes for pairwise comparisons
Tol_PAM.pair.es_TP1<-data.frame(eff_size(Tol_PAM.pair_TP1, sigma=sigma(Tol_PAM.lme_TP1), edf=df.residual(Tol_PAM.lme_TP1)))
Tol_PAM.pair.es_TP1
Tol_PAM.pair.es_TP1<-Tol_PAM.pair.es_TP1 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_PAM.pair_TP1.res<-merge(data.frame(Tol_PAM.pair_TP1$contrasts), Tol_PAM.pair.es_TP1[,-c(1)])
Tol_PAM.pair_TP1.res$Response<-rep("FvFm Retention", nrow(Tol_PAM.pair_TP1.res))

##Add Significance levels
Tol_PAM.pair_TP1.res$Sig<-ifelse(Tol_PAM.pair_TP1.res$p.value<0.001, "***",
                              ifelse(Tol_PAM.pair_TP1.res$p.value<0.01, "**",
                                     ifelse(Tol_PAM.pair_TP1.res$p.value<0.05, "*",
                                            ifelse(Tol_PAM.pair_TP1.res$p.value<0.1, "-", NA))))
Tol_PAM.pair_TP1.res
```


```{r Plot FvFm Retention TP1}
##Summary statistics by Site and Origin
TP1_TolPAM.sum<-summarySE(Therm_TP1, measurevar="PAM.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average FvFm Retention across Treatments
TP1_TolPAM.plot<-ggplot(TP1_TolPAM.sum, aes(x=Site, y=PAM.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=PAM.prop-se, ymax=PAM.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Photochemical Efficiency Retention", colour="Origin")+
  ylim(0.6, 1)+
  annotate("text", x=1, y=.96, label="*", size=sig.sz, fontface="bold"); TP1_TolPAM.plot
```


#### Save Results
```{r}
##Combine Results
Tol_TP1.res<-rbind(Tol_PAM_TP1.res, Tol_Sym_TP1.res, Tol_Chl_TP1.res) 
Tol.pair.res_TP1<-rbind(Tol_PAM.pair_TP1.res, Tol_Sym.pair_TP1.res, Tol_Chl.pair_TP1.res)

##Add Timepoint
Tol_TP1.res$TimeP<-rep("TP1", nrow(Tol_TP1.res))
Tol.pair.res_TP1$TimeP<-rep("TP1", nrow(Tol.pair.res_TP1))

```


### TP2
```{r}
##Subset Timepoint 2
Therm_TP2<-subset(Therm_H, TimeP=="TP2")
```


#### Symbionts
```{r}
#Check normality
hist(Therm_TP2$Sym.prop)
shapiro.test(Therm_TP2$Sym.prop)
#Not normal

hist(Therm_TP2$Sym.prop^2)
shapiro.test(Therm_TP2$Sym.prop^2)
##Still not normal

##Model with no transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Sym.lme_TP2<-lmer(Sym.prop~Origin*Site+(1|Genotype), data=Therm_TP2)

##Check residuals
Tol_Sym.lme_res_TP2 <- simulateResiduals(fittedModel = Tol_Sym.lme_TP2, plot = F)
plot(Tol_Sym.lme_res_TP2)

##Model results
summary(Tol_Sym.lme_TP2)

eta_squared(Tol_Sym.lme_TP2)

##Save model results
Tol_Sym_TP2.res<-data.frame(summary(Tol_Sym.lme_TP2)$coefficients[-1,])
Tol_Sym_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Sym_TP2.res$EtaSq<-c(eta_squared(Tol_Sym.lme_TP2)$Eta2)
Tol_Sym_TP2.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym_TP2.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Sym.pair_TP2<-emmeans(Tol_Sym.lme_TP2, pairwise~Origin | Site)
Tol_Sym.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Tol_Sym.pair.es_TP2<-data.frame(eff_size(Tol_Sym.pair_TP2, sigma=sigma(Tol_Sym.lme_TP2), edf=df.residual(Tol_Sym.lme_TP2)))
Tol_Sym.pair.es_TP2
Tol_Sym.pair.es_TP2<-Tol_Sym.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Sym.pair_TP2.res<-merge(data.frame(Tol_Sym.pair_TP2$contrasts), Tol_Sym.pair.es_TP2[,-c(1)])
Tol_Sym.pair_TP2.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym.pair_TP2.res))

##Add Significance levels
Tol_Sym.pair_TP2.res$Sig<-ifelse(Tol_Sym.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Tol_Sym.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Tol_Sym.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Tol_Sym.pair_TP2.res$p.value<0.1, "-", NA))))
Tol_Sym.pair_TP2.res
```


```{r Plot Symbiont Retention TP2}
##Summary statistics by Site and Origin
TP2_TolSym.sum<-summarySE(Therm_TP2, measurevar="Sym.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP2_TolSym.plot<-ggplot(TP2_TolSym.sum, aes(x=Site, y=Sym.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Sym.prop-se, ymax=Sym.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Symbiont Retention", colour="Origin")+
  ylim(0, 1)+
 geom_bracket(xmin=1, xmax=2, y.position=.04, label.size=levels.sz, label="*", inherit.aes = FALSE); TP2_TolSym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Therm_TP2$Chl.prop)
shapiro.test(Therm_TP2$Chl.prop)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Chl.lme_TP2<-lmer(Chl.prop~Origin*Site+(1|Genotype), data=Therm_TP2)

##Check residuals
Tol_Chl.lme_res_TP2 <- simulateResiduals(fittedModel = Tol_Chl.lme_TP2, plot = F)
plot(Tol_Chl.lme_res_TP2)

##Model results
summary(Tol_Chl.lme_TP2)

eta_squared(Tol_Chl.lme_TP2)

##Save model results
Tol_Chl_TP2.res<-data.frame(summary(Tol_Chl.lme_TP2)$coefficients[-1,])
Tol_Chl_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Chl_TP2.res$EtaSq<-c(eta_squared(Tol_Chl.lme_TP2)$Eta2)
Tol_Chl_TP2.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl_TP2.res))

```


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Chl.pair_TP2<-emmeans(Tol_Chl.lme_TP2, pairwise~Origin | Site)
Tol_Chl.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Tol_Chl.pair.es_TP2<-data.frame(eff_size(Tol_Chl.pair_TP2, sigma=sigma(Tol_Chl.lme_TP2), edf=df.residual(Tol_Chl.lme_TP2)))
Tol_Chl.pair.es_TP2
Tol_Chl.pair.es_TP2<-Tol_Chl.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Chl.pair_TP2.res<-merge(data.frame(Tol_Chl.pair_TP2$contrasts), Tol_Chl.pair.es_TP2[,-c(1)])
Tol_Chl.pair_TP2.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl.pair_TP2.res))

##Add Significance levels
Tol_Chl.pair_TP2.res$Sig<-ifelse(Tol_Chl.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Tol_Chl.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Tol_Chl.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Tol_Chl.pair_TP2.res$p.value<0.1, "-", NA))))
Tol_Chl.pair_TP2.res
```


```{r Plot Chlorophyll Retention TP2}
##Summary statistics by Site and Origin
TP2_TolChl.sum<-summarySE(Therm_TP2, measurevar="Chl.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP2_TolChl.plot<-ggplot(TP2_TolChl.sum, aes(x=Site, y=Chl.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl.prop-se, ymax=Chl.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Chlorophyll Retention", colour="Origin")+
  ylim(0, 0.35)+
  geom_bracket(xmin=1, xmax=2, y.position=.02, label.size=levels.sz, label="-", inherit.aes = FALSE); TP2_TolChl.plot
```


#### FvFm
```{r}
#Check normality
hist(Therm_TP2$PAM.prop)
shapiro.test(Therm_TP2$PAM.prop)
#Not normal

hist(Therm_TP2$PAM.prop^2)
shapiro.test(Therm_TP2$PAM.prop^2)
##Still not normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_PAM.lme_TP2<-lmer(PAM.prop~Origin*Site+(1|Genotype), data=Therm_TP2)

##Check residuals
Tol_PAM.lme_res_TP2 <- simulateResiduals(fittedModel = Tol_PAM.lme_TP2, plot = F)
plot(Tol_PAM.lme_res_TP2)

##Model results
summary(Tol_PAM.lme_TP2)

eta_squared(Tol_PAM.lme_TP2)

##Save model results
Tol_PAM_TP2.res<-data.frame(summary(Tol_PAM.lme_TP2)$coefficients[-1,])
Tol_PAM_TP2.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_PAM_TP2.res$EtaSq<-c(eta_squared(Tol_PAM.lme_TP2)$Eta2)
Tol_PAM_TP2.res$Response<-rep("FvFm Retention", nrow(Tol_PAM_TP2.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_PAM.pair_TP2<-emmeans(Tol_PAM.lme_TP2, pairwise~Origin | Site)
Tol_PAM.pair_TP2

#Calculate standardized effect sizes for pairwise comparisons
Tol_PAM.pair.es_TP2<-data.frame(eff_size(Tol_PAM.pair_TP2, sigma=sigma(Tol_PAM.lme_TP2), edf=df.residual(Tol_PAM.lme_TP2)))
Tol_PAM.pair.es_TP2
Tol_PAM.pair.es_TP2<-Tol_PAM.pair.es_TP2 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_PAM.pair_TP2.res<-merge(data.frame(Tol_PAM.pair_TP2$contrasts), Tol_PAM.pair.es_TP2[,-c(1)])
Tol_PAM.pair_TP2.res$Response<-rep("FvFm Retention", nrow(Tol_PAM.pair_TP2.res))

##Add Significance levels
Tol_PAM.pair_TP2.res$Sig<-ifelse(Tol_PAM.pair_TP2.res$p.value<0.001, "***",
                              ifelse(Tol_PAM.pair_TP2.res$p.value<0.01, "**",
                                     ifelse(Tol_PAM.pair_TP2.res$p.value<0.05, "*",
                                            ifelse(Tol_PAM.pair_TP2.res$p.value<0.1, "-", NA))))
Tol_PAM.pair_TP2.res
```


```{r Plot FvFm Retention TP2}
##Summary statistics by Site and Origin
TP2_TolPAM.sum<-summarySE(Therm_TP2, measurevar="PAM.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP2_TolPAM.plot<-ggplot(TP2_TolPAM.sum, aes(x=Site, y=PAM.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=PAM.prop-se, ymax=PAM.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Photochemical Efficiency Retention", colour="Origin")+
  ylim(0.6, 1)+
  annotate("text", x=1, y=0.96, label="**", size=sig.sz, fontface="bold")+
 geom_bracket(xmin=1, xmax=2, y.position=.612, label.size=levels.sz, label="*", inherit.aes = FALSE); TP2_TolPAM.plot
```


#### Save Results
```{r}
##Combine Results
Tol_TP2.res<-rbind(Tol_PAM_TP2.res, Tol_Sym_TP2.res, Tol_Chl_TP2.res)
Tol.pair.res_TP2<-rbind(Tol_PAM.pair_TP2.res, Tol_Sym.pair_TP2.res, Tol_Chl.pair_TP2.res)

##Add Timepoint
Tol_TP2.res$TimeP<-rep("TP2", nrow(Tol_TP2.res))
Tol.pair.res_TP2$TimeP<-rep("TP2", nrow(Tol.pair.res_TP2))

```


### TP3
```{r}
##Subset Timepoint 3
Therm_TP3<-subset(Therm_H, TimeP=="TP3")
```


#### Symbionts
```{r}
#Check normality
hist(Therm_TP3$Sym.prop)
shapiro.test(Therm_TP3$Sym.prop)
#Not normal

hist(Therm_TP3$Sym.prop^2)
shapiro.test(Therm_TP3$Sym.prop^2)
##Still not normal but less skewed

##Model with no transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Sym.lme_TP3<-lmer(Sym.prop~Origin*Site+(1|Genotype), data=Therm_TP3)

##Check residuals
Tol_Sym.lme_res_TP3 <- simulateResiduals(fittedModel = Tol_Sym.lme_TP3, plot = F)
plot(Tol_Sym.lme_res_TP3)

##Model results
summary(Tol_Sym.lme_TP3)

eta_squared(Tol_Sym.lme_TP3)

##Save model results
Tol_Sym_TP3.res<-data.frame(summary(Tol_Sym.lme_TP3)$coefficients[-1,])
Tol_Sym_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Sym_TP3.res$EtaSq<-c(eta_squared(Tol_Sym.lme_TP3)$Eta2)
Tol_Sym_TP3.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym_TP3.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Sym.pair_TP3<-emmeans(Tol_Sym.lme_TP3, pairwise~Origin | Site)
Tol_Sym.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Tol_Sym.pair.es_TP3<-data.frame(eff_size(Tol_Sym.pair_TP3, sigma=sigma(Tol_Sym.lme_TP3), edf=df.residual(Tol_Sym.lme_TP3)))
Tol_Sym.pair.es_TP3
Tol_Sym.pair.es_TP3<-Tol_Sym.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Sym.pair_TP3.res<-merge(data.frame(Tol_Sym.pair_TP3$contrasts), Tol_Sym.pair.es_TP3[,-c(1)])
Tol_Sym.pair_TP3.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym.pair_TP3.res))

##Add Significance levels
Tol_Sym.pair_TP3.res$Sig<-ifelse(Tol_Sym.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Tol_Sym.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Tol_Sym.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Tol_Sym.pair_TP3.res$p.value<0.1, "-", NA))))
Tol_Sym.pair_TP3.res
```


```{r Plot Symbiont Retention TP3}
##Summary statistics by Site and Origin
TP3_TolSym.sum<-summarySE(Therm_TP3, measurevar="Sym.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP3_TolSym.plot<-ggplot(TP3_TolSym.sum, aes(x=Site, y=Sym.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Sym.prop-se, ymax=Sym.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Symbiont Retention", colour="Origin")+
  ylim(0, 1)+
 geom_bracket(xmin=1, xmax=2, y.position=.04, label.size=levels.sz, label="*", inherit.aes = FALSE); TP3_TolSym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Therm_TP3$Chl.prop)
shapiro.test(Therm_TP3$Chl.prop)
#Not normal

hist(log(Therm_TP3$Chl.prop+1))
shapiro.test(log(Therm_TP3$Chl.prop+1))
##Still not normal but less skewed

##Model with log +1 transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Chl.lme_TP3<-lmer(log(Chl.prop+1)~Origin*Site+(1|Genotype), data=Therm_TP3)

##Check residuals
Tol_Chl.lme_res_TP3 <- simulateResiduals(fittedModel = Tol_Chl.lme_TP3, plot = F)
plot(Tol_Chl.lme_res_TP3)

##Model results
summary(Tol_Chl.lme_TP3)

eta_squared(Tol_Chl.lme_TP3)

##Save model results
Tol_Chl_TP3.res<-data.frame(summary(Tol_Chl.lme_TP3)$coefficients[-1,])
Tol_Chl_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Chl_TP3.res$EtaSq<-c(eta_squared(Tol_Chl.lme_TP3)$Eta2)
Tol_Chl_TP3.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl_TP3.res))

```
Not normal, but model residuals are OK.


```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Chl.pair_TP3<-emmeans(Tol_Chl.lme_TP3, pairwise~Origin | Site)
Tol_Chl.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Tol_Chl.pair.es_TP3<-data.frame(eff_size(Tol_Chl.pair_TP3, sigma=sigma(Tol_Chl.lme_TP3), edf=df.residual(Tol_Chl.lme_TP3)))
Tol_Chl.pair.es_TP3
Tol_Chl.pair.es_TP3<-Tol_Chl.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Chl.pair_TP3.res<-merge(data.frame(Tol_Chl.pair_TP3$contrasts), Tol_Chl.pair.es_TP3[,-c(1)])
Tol_Chl.pair_TP3.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl.pair_TP3.res))

##Add Significance levels
Tol_Chl.pair_TP3.res$Sig<-ifelse(Tol_Chl.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Tol_Chl.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Tol_Chl.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Tol_Chl.pair_TP3.res$p.value<0.1, "-", NA))))
Tol_Chl.pair_TP3.res
```


```{r Plot Chlorophyll Retention TP3}
##Summary statistics by Site and Origin
TP3_TolChl.sum<-summarySE(Therm_TP3, measurevar="Chl.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP3_TolChl.plot<-ggplot(TP3_TolChl.sum, aes(x=Site, y=Chl.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl.prop-se, ymax=Chl.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Chlorophyll Retention", colour="Origin")+
  ylim(0, 1); TP3_TolChl.plot
```


#### FvFm
```{r}
#Check normality
hist(Therm_TP3$PAM.prop)
shapiro.test(Therm_TP3$PAM.prop)
#Normal

##Model
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_PAM.lme_TP3<-lmer(PAM.prop~Origin*Site+(1|Genotype), data=Therm_TP3)

##Check residuals
Tol_PAM.lme_res_TP3 <- simulateResiduals(fittedModel = Tol_PAM.lme_TP3, plot = F)
plot(Tol_PAM.lme_res_TP3)

##Model results
summary(Tol_PAM.lme_TP3)

eta_squared(Tol_PAM.lme_TP3)

##Save model results
Tol_PAM_TP3.res<-data.frame(summary(Tol_PAM.lme_TP3)$coefficients[-1,])
Tol_PAM_TP3.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_PAM_TP3.res$EtaSq<-c(eta_squared(Tol_PAM.lme_TP3)$Eta2)
Tol_PAM_TP3.res$Response<-rep("FvFm Retention", nrow(Tol_PAM_TP3.res))

```

```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_PAM.pair_TP3<-emmeans(Tol_PAM.lme_TP3, pairwise~Origin | Site)
Tol_PAM.pair_TP3

#Calculate standardized effect sizes for pairwise comparisons
Tol_PAM.pair.es_TP3<-data.frame(eff_size(Tol_PAM.pair_TP3, sigma=sigma(Tol_PAM.lme_TP3), edf=df.residual(Tol_PAM.lme_TP3)))
Tol_PAM.pair.es_TP3
Tol_PAM.pair.es_TP3<-Tol_PAM.pair.es_TP3 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_PAM.pair_TP3.res<-merge(data.frame(Tol_PAM.pair_TP3$contrasts), Tol_PAM.pair.es_TP3[,-c(1)])
Tol_PAM.pair_TP3.res$Response<-rep("FvFm Retention", nrow(Tol_PAM.pair_TP3.res))

##Add Significance levels
Tol_PAM.pair_TP3.res$Sig<-ifelse(Tol_PAM.pair_TP3.res$p.value<0.001, "***",
                              ifelse(Tol_PAM.pair_TP3.res$p.value<0.01, "**",
                                     ifelse(Tol_PAM.pair_TP3.res$p.value<0.05, "*",
                                            ifelse(Tol_PAM.pair_TP3.res$p.value<0.1, "-", NA))))
Tol_PAM.pair_TP3.res
```


```{r Plot FvFm Retention TP3}
##Summary statistics by Site and Origin
TP3_TolPAM.sum<-summarySE(Therm_TP3, measurevar="PAM.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP3_TolPAM.plot<-ggplot(TP3_TolPAM.sum, aes(x=Site, y=PAM.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=PAM.prop-se, ymax=PAM.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Photochemical Efficiency Retention", colour="Origin")+
  ylim(0.7, 1); TP3_TolPAM.plot
```


#### Save Results
```{r}
##Combine Results
Tol_TP3.res<-rbind(Tol_PAM_TP3.res, Tol_Sym_TP3.res, Tol_Chl_TP3.res)
Tol.pair.res_TP3<-rbind(Tol_PAM.pair_TP3.res, Tol_Sym.pair_TP3.res, Tol_Chl.pair_TP3.res)

##Add Timepoint
Tol_TP3.res$TimeP<-rep("TP3", nrow(Tol_TP3.res))
Tol.pair.res_TP3$TimeP<-rep("TP3", nrow(Tol.pair.res_TP3))

```


### TP4
```{r}
##Subset Timepoint 4
Therm_TP4<-subset(Therm_H, TimeP=="TP4")
```


#### Symbionts
```{r}
#Check normality
hist(Therm_TP4$Sym.prop)
shapiro.test(Therm_TP4$Sym.prop)
#Not normal

hist(log(Therm_TP4$Sym.prop+1))
shapiro.test(log(Therm_TP4$Sym.prop+1))
#Normal

##Model with log +1 transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Sym.lme_TP4<-lmer(log(Sym.prop+1)~Origin*Site+(1|Genotype), data=Therm_TP4)

##Check residuals
Tol_Sym.lme_res_TP4 <- simulateResiduals(fittedModel = Tol_Sym.lme_TP4, plot = F)
plot(Tol_Sym.lme_res_TP4)

##Model results
summary(Tol_Sym.lme_TP4)

eta_squared(Tol_Sym.lme_TP4)

##Save model results
Tol_Sym_TP4.res<-data.frame(summary(Tol_Sym.lme_TP4)$coefficients[-1,])
Tol_Sym_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Sym_TP4.res$EtaSq<-c(eta_squared(Tol_Sym.lme_TP4)$Eta2)
Tol_Sym_TP4.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym_TP4.res))

```

```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Sym.pair_TP4<-emmeans(Tol_Sym.lme_TP4, pairwise~Origin | Site)
Tol_Sym.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Tol_Sym.pair.es_TP4<-data.frame(eff_size(Tol_Sym.pair_TP4, sigma=sigma(Tol_Sym.lme_TP4), edf=df.residual(Tol_Sym.lme_TP4)))
Tol_Sym.pair.es_TP4
Tol_Sym.pair.es_TP4<-Tol_Sym.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Sym.pair_TP4.res<-merge(data.frame(Tol_Sym.pair_TP4$contrasts), Tol_Sym.pair.es_TP4[,-c(1)])
Tol_Sym.pair_TP4.res$Response<-rep("Symbiont Retention", nrow(Tol_Sym.pair_TP4.res))

##Add Significance levels
Tol_Sym.pair_TP4.res$Sig<-ifelse(Tol_Sym.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Tol_Sym.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Tol_Sym.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Tol_Sym.pair_TP4.res$p.value<0.1, "-", NA))))
Tol_Sym.pair_TP4.res
```


```{r Plot Symbiont Retention TP4}
##Summary statistics by Site and Origin
TP4_TolSym.sum<-summarySE(Therm_TP4, measurevar="Sym.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP4_TolSym.plot<-ggplot(TP4_TolSym.sum, aes(x=Site, y=Sym.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Sym.prop-se, ymax=Sym.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Symbiont Retention", colour="Origin")+
  ylim(0, 1); TP4_TolSym.plot
```


#### Chlorophyll
```{r}
#Check normality
hist(Therm_TP4$Chl.prop)
shapiro.test(Therm_TP4$Chl.prop)
#Not normal

hist(log(Therm_TP4$Chl.prop+1))
shapiro.test(log(Therm_TP4$Chl.prop+1))
##Nearly normal

##Model with log +1 transformation
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_Chl.lme_TP4<-lmer(log(Chl.prop+1)~Origin*Site+(1|Genotype), data=Therm_TP4)

##Check residuals
Tol_Chl.lme_res_TP4 <- simulateResiduals(fittedModel = Tol_Chl.lme_TP4, plot = F)
plot(Tol_Chl.lme_res_TP4)

##Model results
summary(Tol_Chl.lme_TP4)

eta_squared(Tol_Chl.lme_TP4)

##Save model results
Tol_Chl_TP4.res<-data.frame(summary(Tol_Chl.lme_TP4)$coefficients[-1,])
Tol_Chl_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_Chl_TP4.res$EtaSq<-c(eta_squared(Tol_Chl.lme_TP4)$Eta2)
Tol_Chl_TP4.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl_TP4.res))

```

```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_Chl.pair_TP4<-emmeans(Tol_Chl.lme_TP4, pairwise~Origin | Site)
Tol_Chl.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Tol_Chl.pair.es_TP4<-data.frame(eff_size(Tol_Chl.pair_TP4, sigma=sigma(Tol_Chl.lme_TP4), edf=df.residual(Tol_Chl.lme_TP4)))
Tol_Chl.pair.es_TP4
Tol_Chl.pair.es_TP4<-Tol_Chl.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_Chl.pair_TP4.res<-merge(data.frame(Tol_Chl.pair_TP4$contrasts), Tol_Chl.pair.es_TP4[,-c(1)])
Tol_Chl.pair_TP4.res$Response<-rep("Chlorophyll Retention", nrow(Tol_Chl.pair_TP4.res))

##Add Significance levels
Tol_Chl.pair_TP4.res$Sig<-ifelse(Tol_Chl.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Tol_Chl.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Tol_Chl.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Tol_Chl.pair_TP4.res$p.value<0.1, "-", NA))))
Tol_Chl.pair_TP4.res
```


```{r Plot Chlorophyll Retention TP4}
##Summary statistics by Site and Origin
TP4_TolChl.sum<-summarySE(Therm_TP4, measurevar="Chl.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP4_TolChl.plot<-ggplot(TP4_TolChl.sum, aes(x=Site, y=Chl.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=Chl.prop-se, ymax=Chl.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Chlorophyll Retention", colour="Origin")+
  ylim(0, 0.35); TP4_TolChl.plot
```


#### FvFm
```{r}
#Check normality
hist(Therm_TP4$PAM.prop)
shapiro.test(Therm_TP4$PAM.prop)
#Not normal

hist(Therm_TP4$PAM.prop^2)
shapiro.test(Therm_TP4$PAM.prop^2)
##Still not normal

##Model with squared transformation to improve residuals
#Function of Site and Origin, with Genotype as a Random effect
#Interaction between Site and Origin
Tol_PAM.lme_TP4<-lmer(PAM.prop^2~Origin*Site+(1|Genotype), data=Therm_TP4)

##Check residuals
Tol_PAM.lme_res_TP4 <- simulateResiduals(fittedModel = Tol_PAM.lme_TP4, plot = F)
plot(Tol_PAM.lme_res_TP4)

##Model results
summary(Tol_PAM.lme_TP4)

eta_squared(Tol_PAM.lme_TP4)

##Save model results
Tol_PAM_TP4.res<-data.frame(summary(Tol_PAM.lme_TP4)$coefficients[-1,])
Tol_PAM_TP4.res$Predictor<-c("Origin", "Site", "Origin x Site")
Tol_PAM_TP4.res$EtaSq<-c(eta_squared(Tol_PAM.lme_TP4)$Eta2)
Tol_PAM_TP4.res$Response<-rep("FvFm Retention", nrow(Tol_PAM_TP4.res))

```

```{r}
##Pairwise Comparisons of Interest
#Native vs Transplant within a Site
Tol_PAM.pair_TP4<-emmeans(Tol_PAM.lme_TP4, pairwise~Origin | Site)
Tol_PAM.pair_TP4

#Calculate standardized effect sizes for pairwise comparisons
Tol_PAM.pair.es_TP4<-data.frame(eff_size(Tol_PAM.pair_TP4, sigma=sigma(Tol_PAM.lme_TP4), edf=df.residual(Tol_PAM.lme_TP4)))
Tol_PAM.pair.es_TP4
Tol_PAM.pair.es_TP4<-Tol_PAM.pair.es_TP4 %>% dplyr::rename(contrast.se = contrast, SE.es = SE, df.es = df)

##Save Results
Tol_PAM.pair_TP4.res<-merge(data.frame(Tol_PAM.pair_TP4$contrasts), Tol_PAM.pair.es_TP4[,-c(1)])
Tol_PAM.pair_TP4.res$Response<-rep("FvFm Retention", nrow(Tol_PAM.pair_TP4.res))

##Add Significance levels
Tol_PAM.pair_TP4.res$Sig<-ifelse(Tol_PAM.pair_TP4.res$p.value<0.001, "***",
                              ifelse(Tol_PAM.pair_TP4.res$p.value<0.01, "**",
                                     ifelse(Tol_PAM.pair_TP4.res$p.value<0.05, "*",
                                            ifelse(Tol_PAM.pair_TP4.res$p.value<0.1, "-", NA))))
Tol_PAM.pair_TP4.res
```


```{r Plot FvFm Retention TP4}
##Summary statistics by Site and Origin
TP4_TolPAM.sum<-summarySE(Therm_TP4, measurevar="PAM.prop", groupvars=c("Site", "Origin", "Site.Orig"), na.rm=TRUE)

##Plot Average Retention across Treatments
TP4_TolPAM.plot<-ggplot(TP4_TolPAM.sum, aes(x=Site, y=PAM.prop, colour=Site.Orig)) + 
  scale_colour_manual(values=Orig.colors.o)+
  geom_errorbar(aes(ymin=PAM.prop-se, ymax=PAM.prop+se), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(size=point.sz, position=position_dodge(width=0.5))+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x="Site and Origin", y="Photochemical Efficiency Retention", colour="Origin")+
  ylim(0.6, 1); TP4_TolPAM.plot
```


#### Save Results
```{r}
##Combine Results
Tol_TP4.res<-rbind(Tol_PAM_TP4.res, Tol_Sym_TP4.res, Tol_Chl_TP4.res)
Tol.pair.res_TP4<-rbind(Tol_PAM.pair_TP4.res, Tol_Sym.pair_TP4.res, Tol_Chl.pair_TP4.res)

##Add Timepoint
Tol_TP4.res$TimeP<-rep("TP4", nrow(Tol_TP4.res))
Tol.pair.res_TP4$TimeP<-rep("TP4", nrow(Tol.pair.res_TP4))

```


### Percent Difference in Fv/Fm
Focus on Fv/Fm at KL TP 1 and 2 where difference were significant

```{r}
##Summary statistics by Genotype, Origin, Site, and Timepoint
Tol.PAM.sum<-summarySE(Therm_H, measurevar="PAM.prop", groupvars=c("Genotype", "Origin", "Site", "TimeP"), na.rm=TRUE)

##Remove Initial Timepoint
Tol.PAM.sum<-Tol.PAM.sum[-c(which(Tol.PAM.sum$TimeP=="IN")),]

##Calculate Percent Difference between Native and Transplant by Genotype
Tol.PAM.dif<-Tol.PAM.sum[which(Tol.PAM.sum$Origin=="Native"), c(1,3,4,6)]
Tol.PAM.dif<-Tol.PAM.dif %>% dplyr::rename(PAM_N = PAM.prop)
Tol.PAM.dif$PAM_T<-Tol.PAM.sum$PAM.prop[which(Tol.PAM.sum$Origin=="Transplant")]

##KL
Tol.PAM.dif_KL<-subset(Tol.PAM.dif, Site=="KL")
Tol.PAM.dif_KL$Perc.Inc<-((Tol.PAM.dif_KL$PAM_T-Tol.PAM.dif_KL$PAM_N)/Tol.PAM.dif_KL$PAM_N)*100
Tol.PAM.dif_KL

##Mean and Standard Error by Timepoint
Tol.PAM.dif_KL.sum<-summarySE(Tol.PAM.dif_KL, measurevar="Perc.Inc", groupvars=c("TimeP"), na.rm=TRUE)
Tol.PAM.dif_KL.sum
```


# Performance Effect Sizes

### Current vs Original Site
```{r}
##Combine Results
Growth.TP.res<-rbind(TLE_TP1.2.res, TLE_TP3.4.res)
Growth.TP.res<-Growth.TP.res %>% dplyr::rename(TimeP = Timepoint)
Tol_TP.res<-rbind(Tol_TP1.res, Tol_TP2.res, Tol_TP3.res, Tol_TP4.res)
Perf.res<-rbind(Growth.TP.res, Tol_TP.res)
Perf.res<-Perf.res %>% dplyr::rename(p = Pr...t..)

##Responses for Figure
Perf.res.plot<-Perf.res[-c(which(Perf.res$Response=="Chlorophyll Retention")),]

##Add Significance levels
Perf.res.plot$Sig<-ifelse(Perf.res.plot$p<0.001, "***", ifelse(Perf.res.plot$p<0.01, "**", 
            ifelse(Perf.res.plot$p<0.05, "*", ifelse(Perf.res.plot$p<0.1, "-", NA))))

##Set Order
Perf.res.plot$Response<-factor(Perf.res.plot$Response, levels=c("Growth", "FvFm Retention", "Symbiont Retention"), ordered=TRUE)

##Match Growth Timepoints
Perf.res.plot$TimeP[which(Perf.res.plot$TimeP=="TP1_2")]<-"TP2"
Perf.res.plot$TimeP[which(Perf.res.plot$TimeP=="TP3_4")]<-"TP4"

```


#### Plot Effect Size of Origin and Site
```{r Plot Origin and Site Effect Sizes}
Perf.Site.Orig.ES.plot<-ggplot(Perf.res.plot, aes(fill=Predictor, y=EtaSq, x=TimeP)) + 
    geom_bar(position="stack", stat="identity")+
  facet_wrap(vars(Response))+
  theme_bw()+
  scale_fill_manual(values=c("indianred2", "mediumpurple3", "steelblue2"))+
  theme(strip.text=element_text(size=axis.title.sz),
        strip.background = element_rect(fill="white"),
        axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position=c(.08, .8), 
        panel.grid.major.x=element_blank())+
   labs(x="Time Point", y=expression(paste("Predictor Effect Size (p", eta^2, ")")), colour="Predictor")+
   geom_text(aes(label=Sig), position=position_stack(vjust=0.5), size=sig.sz); Perf.Site.Orig.ES.plot
```


### Native vs Transplant
```{r}
##Combine Pairwise Results
Growth.pair.res<-rbind(TLE.pair_TP1.2.res, TLE.pair_TP3.4.res)
Tol.pair.res<-rbind(Tol.pair.res_TP1, Tol.pair.res_TP2, Tol.pair.res_TP3, Tol.pair.res_TP4)
Perf.pair.res<-rbind(Growth.pair.res[,c(1:13, 15, 14)], Tol.pair.res)

##Responses for Figure
Perf.pair.res.plot<-Perf.pair.res[-c(which(Perf.pair.res$Response=="Chlorophyll Retention")),]

##Add Significance
Perf.pair.res.plot$Significant<-ifelse(Perf.pair.res.plot$p.value<0.05, "sig", "non")

##Set Order
Perf.pair.res.plot$Response<-factor(Perf.pair.res.plot$Response, levels=c("Growth", "FvFm Retention", "Symbiont Retention"), ordered=TRUE)

##Match Growth Timepoints
Perf.pair.res.plot$TimeP[which(Perf.pair.res.plot$TimeP=="TP1_2")]<-"TP2"
Perf.pair.res.plot$TimeP[which(Perf.pair.res.plot$TimeP=="TP3_4")]<-"TP4"

##Set Factors
Perf.pair.res.plot$Site<-factor(Perf.pair.res.plot$Site, levels=c("KL", "SS"), ordered=TRUE)

Perf.pair.res.plot$TimeP<-factor(Perf.pair.res.plot$TimeP, levels=c("TP1", "TP2", "TP3", "TP4"), ordered=TRUE)
```


#### Plot Effect Size of Origin
```{r Plot Origin Effect Sizes}
##Plot Effect Size across Timepoints
Perf.Orig.ES.plot<-ggplot(Perf.pair.res.plot, aes(x=TimeP, y=effect.size, group=Site)) +
  geom_rect(fill="grey95", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.8, ymax=0.8)+
  geom_rect(fill="grey85", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.5, ymax=0.5)+
  geom_rect(fill="grey70", alpha=0.4, xmin=-Inf, xmax=Inf, ymin=-.2, ymax=0.2)+
  geom_hline(yintercept=0, linetype="dashed", color="grey50", linewidth=0.5)+
  geom_errorbar(aes(ymin=effect.size-SE.es, ymax=effect.size+SE.es, colour=Site, group=Site), width=cap.sz, linewidth=bar.sz, position=position_dodge(width=0.5)) +
  geom_point(aes(shape=Significant, colour=Site), size=point.sz, position=position_dodge(width=0.5))+
  facet_wrap(vars(Response))+
  scale_colour_manual(values=Site.colors.o)+
  scale_shape_manual(values=c(1, 16))+
  theme_bw()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), 
        legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"), 
        legend.position=c(.1, .1),
        legend.direction = "horizontal",
        strip.text=element_text(size=axis.title.sz),
        strip.background = element_rect(fill="white"), 
        panel.grid.major.y=element_blank())+
  labs(x="Time Point", y="Origin Effect Size (Cohen's d)", colour="Site")+
  guides(shape="none"); Perf.Orig.ES.plot
```
Need to fix dodge for points and error bars


# Trade Offs

### Initial T1 and T2

#### Average Performance
Calculating the average of Growth (TP1 to TP2) and Thermal Tolerance (Retention of Symbionts, Chlorophyll, and Photochemical Efficiency at TP1) for each Set (Site, Genotype, Origin).

#### Growth
```{r}
##Average
names(Growth_TP1.2)
TP1.2_Growth_Set<-aggregate(Growth_TP1.2$TLE_cm.day, list(Growth_TP1.2$Site, Growth_TP1.2$Genotype, Growth_TP1.2$Orig), mean)
names(TP1.2_Growth_Set)<-c("Site", "Genotype", "Orig", "TLE_cm.day")

##Add Standard Error
TP1.2_Growth_Set$TLE_se<-aggregate(Growth_TP1.2$TLE_cm.day, list(Growth_TP1.2$Site, Growth_TP1.2$Genotype, Growth_TP1.2$Orig), std.error)[,4]
```


#### Tolerance
```{r}
##Average
names(Therm_TP1)
T1_Therm_Set<-aggregate(Therm_TP1[,c(19:21)], list(Therm_TP1$Site, Therm_TP1$Genotype, Therm_TP1$Orig), mean)
names(T1_Therm_Set)<-c("Site", "Genotype", "Orig", "Chl.prop", "Sym.prop", "PAM.prop")

##Add Standard Error
T1_Therm_Set$Chl_se<-aggregate(Therm_TP1$Chl.prop, list(Therm_TP1$Site, Therm_TP1$Genotype, Therm_TP1$Orig), std.error)[,4]

T1_Therm_Set$Sym_se<-aggregate(Therm_TP1$Sym.prop, list(Therm_TP1$Site, Therm_TP1$Genotype, Therm_TP1$Orig), std.error)[,4]

T1_Therm_Set$PAM_se<-aggregate(Therm_TP1$PAM.prop, list(Therm_TP1$Site, Therm_TP1$Genotype, Therm_TP1$Orig), std.error)[,4]
```


#### Merge Growth and Tolerance 
```{r}
TP1.2_Perf_Set<-merge(TP1.2_Growth_Set, T1_Therm_Set, all=TRUE)
TP1.2_Perf_Set$Set<-paste(TP1.2_Perf_Set$Site, TP1.2_Perf_Set$Genotype, TP1.2_Perf_Set$Orig, sep=".")

##Set factor variables
TP1.2_Perf_Set$Site<-factor(TP1.2_Perf_Set$Site, levels=c("KL", "SS"))
TP1.2_Perf_Set$Genotype<-factor(TP1.2_Perf_Set$Genotype, levels=c("AC8", "AC10", "AC12"))
TP1.2_Perf_Set$Orig<-factor(TP1.2_Perf_Set$Orig, levels=c("N", "T"))

##Add a Sample Set Variable
TP1.2_Perf_Set$Set<-paste(TP1.2_Perf_Set$Site, TP1.2_Perf_Set$Genotype, TP1.2_Perf_Set$Orig, sep=".")
TP1.2_Perf_Set$Set<-factor(TP1.2_Perf_Set$Set)

##Add Site.Orig variable
TP1.2_Perf_Set$Site.Orig<-paste(TP1.2_Perf_Set$Site, TP1.2_Perf_Set$Orig, sep=".")
TP1.2_Perf_Set$Site.Orig<-factor(TP1.2_Perf_Set$Site.Orig, levels=c("KL.N", "KL.T", "SS.N", "SS.T"))

```


### Correlation between Growth and Tolerance

#### Symbiont Retention

```{r}
cor.test(TP1.2_Perf_Set$TLE_cm.day, TP1.2_Perf_Set$Sym.prop, method="spearman")
```


```{r Plot Growth vs Symbiont Retention T12}
TP1.2_TLE_Sym.plot<-ggplot(TP1.2_Perf_Set, aes(x=TLE_cm.day, y=Sym.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Symbiont Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = -.15, label=expression(bolditalic(paste(r[S], " = -0.615, p = 0.037"))), size=sig.sz, hjust = 0); TP1.2_TLE_Sym.plot
```


#### Chlorophyll Retention

```{r}
cor.test(TP1.2_Perf_Set$TLE_cm.day, TP1.2_Perf_Set$Chl.prop, method="spearman")
```


```{r Plot Growth vs Chlorophyll Retention T12}
TP1.2_TLE_Chl.plot<-ggplot(TP1.2_Perf_Set, aes(x=TLE_cm.day, y=Chl.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Chlorophyll Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = -.05, label=expression(bolditalic(paste(r[S], " = -0.587, p = 0.049"))), size=sig.sz, hjust = 0); TP1.2_TLE_Chl.plot
```


#### FvFm Retention

```{r}
cor.test(TP1.2_Perf_Set$TLE_cm.day, TP1.2_Perf_Set$PAM.prop, method="spearman")
```


```{r Plot Growth vs FvFm Retention T12}
TP1.2_TLE_PAM.plot<-ggplot(TP1.2_Perf_Set, aes(x=TLE_cm.day, y=PAM.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Photo. Effic. Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = 0.7, label=expression(bolditalic(paste(r[S], " = -0.678, p = 0.019"))), size=sig.sz, hjust = 0); TP1.2_TLE_PAM.plot
```


### Final T3 and T4

#### Average Performance
Calculating the average of Growth (TP3 to TP4) and Thermal Tolerance (Retention of Symbionts, Chlorophyll, and Photochemical Efficiency at TP3) for each Set (Site, Genotype, Origin).

#### Growth
```{r}
##Average
names(Growth_TP3.4)
TP3.4_Growth_Set<-aggregate(Growth_TP3.4$TLE_cm.day, list(Growth_TP3.4$Site, Growth_TP3.4$Genotype, Growth_TP3.4$Orig), mean)
names(TP3.4_Growth_Set)<-c("Site", "Genotype", "Orig", "TLE_cm.day")

##Add Standard Error
TP3.4_Growth_Set$TLE_se<-aggregate(Growth_TP3.4$TLE_cm.day, list(Growth_TP3.4$Site, Growth_TP3.4$Genotype, Growth_TP3.4$Orig), std.error)[,4]
```


#### Tolerance
```{r}
##Average
names(Therm_TP4)
T4_Therm_Set<-aggregate(Therm_TP4[,c(19:21)], list(Therm_TP4$Site, Therm_TP4$Genotype, Therm_TP4$Orig), mean)
names(T4_Therm_Set)<-c("Site", "Genotype", "Orig", "Chl.prop", "Sym.prop", "PAM.prop")

##Add Standard Error
T4_Therm_Set$Chl_se<-aggregate(Therm_TP4$Chl.prop, list(Therm_TP4$Site, Therm_TP4$Genotype, Therm_TP4$Orig), std.error)[,4]

T4_Therm_Set$Sym_se<-aggregate(Therm_TP4$Sym.prop, list(Therm_TP4$Site, Therm_TP4$Genotype, Therm_TP4$Orig), std.error)[,4]

T4_Therm_Set$PAM_se<-aggregate(Therm_TP4$PAM.prop, list(Therm_TP4$Site, Therm_TP4$Genotype, Therm_TP4$Orig), std.error)[,4]
```


#### Merge Growth and Tolerance 
```{r}
TP3.4_Perf_Set<-merge(TP3.4_Growth_Set, T4_Therm_Set, all=TRUE)
TP3.4_Perf_Set$Set<-paste(TP3.4_Perf_Set$Site, TP3.4_Perf_Set$Genotype, TP3.4_Perf_Set$Orig, sep=".")

##Set factor variables
TP3.4_Perf_Set$Site<-factor(TP3.4_Perf_Set$Site, levels=c("KL", "SS"))
TP3.4_Perf_Set$Genotype<-factor(TP3.4_Perf_Set$Genotype, levels=c("AC8", "AC10", "AC12"))
TP3.4_Perf_Set$Orig<-factor(TP3.4_Perf_Set$Orig, levels=c("N", "T"))

##Add a Sample Set Variable
TP3.4_Perf_Set$Set<-paste(TP3.4_Perf_Set$Site, TP3.4_Perf_Set$Genotype, TP3.4_Perf_Set$Orig, sep=".")
TP3.4_Perf_Set$Set<-factor(TP3.4_Perf_Set$Set)

##Add Site.Orig variable
TP3.4_Perf_Set$Site.Orig<-paste(TP3.4_Perf_Set$Site, TP3.4_Perf_Set$Orig, sep=".")
TP3.4_Perf_Set$Site.Orig<-factor(TP3.4_Perf_Set$Site.Orig, levels=c("KL.N", "KL.T", "SS.N", "SS.T"))

```


### Correlation between Growth and Tolerance

#### Symbiont Retention

```{r}
cor.test(TP3.4_Perf_Set$TLE_cm.day, TP3.4_Perf_Set$Sym.prop, method="spearman")
```


```{r Plot Growth vs Symbiont Retention T34}
TP3.4_TLE_Sym.plot<-ggplot(TP3.4_Perf_Set, aes(x=TLE_cm.day, y=Sym.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Symbiont Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = 0.1, label=expression(italic(paste(r[S], " = -0.552, p = 0.067"))), size=sig.sz, hjust = 0); TP3.4_TLE_Sym.plot
```


#### Chlorophyll Retention

```{r}
cor.test(TP3.4_Perf_Set$TLE_cm.day, TP3.4_Perf_Set$Chl.prop, method="spearman")
```


```{r Plot Growth vs Chlorophyll Retention T34}
TP3.4_TLE_Chl.plot<-ggplot(TP3.4_Perf_Set, aes(x=TLE_cm.day, y=Chl.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Chlorophyll Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = 0.065, label=expression(italic(paste(r[S], " = 0.327, p = 0.327"))), size=sig.sz, hjust = 0); TP3.4_TLE_Chl.plot
```


#### FvFm Retention

```{r}
cor.test(TP3.4_Perf_Set$TLE_cm.day, TP3.4_Perf_Set$PAM.prop, method="spearman")
```


```{r Plot Growth vs FvFm Retention T34}
TP3.4_TLE_PAM.plot<-ggplot(TP3.4_Perf_Set, aes(x=TLE_cm.day, y=PAM.prop)) + 
   geom_smooth(method="lm", color="#AA185AFF", fill="#F8D7BFFF", se=TRUE)+
   geom_point(aes(colour=Site.Orig, shape=Genotype),size=point.sz)+
    scale_colour_manual(values=Orig.colors.o)+
  scale_shape_manual(values=Geno.shapes.o)+
   theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"), axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.text=element_text(size=leg.txt.sz), legend.title=element_text(size=leg.title.sz), 
        legend.box.background = element_rect(color = "black"))+
  labs(x=expression(paste('Growth Rate (cm day'^-1*")")), 
       y="Photo. Effic. Retention", 
       colour="Site Origin")+
   annotate("text", x = 0.07, y = 0.5, label=expression(bolditalic(paste(r[S], " = -0.748, p = 0.007"))), size=sig.sz, hjust = 0); TP3.4_TLE_PAM.plot
```


# Figures

### Figure 4 Performance Effect Sizes

#### Adjust for Panel
```{r}
Perf.Site.Orig.ES.plot<-Perf.Site.Orig.ES.plot+ labs(x="")
```


#### Figure 4
```{r}
##Create Panel
Perf.ES_fig<-plot_grid(Perf.Site.Orig.ES.plot, Perf.Orig.ES.plot, 
                    rel_widths=1, rel_heights = 1,
                    nrow=2, ncol=1, byrow=T, labels = c("A", "B"), 
                    label_size=panel.lab.sz)

##Save Figure
ggsave(filename="Figures/03_Performance/Fig4_PerformanceEffectSizes.png", plot=Perf.ES_fig, dpi=300, width=12, height=9, units="in")

```


### Figure 5 Growth vs Tolerance

#### Adjust Figures for Panel
```{r}
##Timepoint 1 to 2
TP1.2_TLE_PAM.plot<-TP1.2_TLE_PAM.plot + ggtitle("Photo. Effic.")+ labs(x=NULL, y = "Tolerance (Retention)")+
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, hjust=0.5), legend.position="none")

TP1.2_TLE_Sym.plot<- TP1.2_TLE_Sym.plot + ggtitle("Symbionts")+ labs(x=NULL, y = NULL)+
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, hjust=0.5)) + ylim(-.15, .8)


##Timepoint 3 to 4
TP3.4_TLE_PAM.plot<-TP3.4_TLE_PAM.plot + labs(y = "Tolerance (Retention)") +
  theme(legend.position="none")

TP3.4_TLE_Sym.plot<-TP3.4_TLE_Sym.plot + labs(y = NULL)

```


#### Figure 5
```{r}
##Create Panel
TradeOff_fig<-plot_grid(TP1.2_TLE_PAM.plot, TP1.2_TLE_Sym.plot,
                      TP3.4_TLE_PAM.plot, TP3.4_TLE_Sym.plot,
                    rel_widths=c(.75, 1), rel_heights = c(1, .98),
                    nrow=2, ncol=2, byrow=T, 
                    labels = c("A", "B", "C", "D"), 
                    label_size=panel.lab.sz)

##Save Figure
ggsave(filename="Figures/03_Performance/Fig5_TradeOff.png", plot=TradeOff_fig, dpi=300, width=8, height=8, units="in")

```


### Figure S5 Growth

#### Adjust Figures for Panel
```{r}
##TP 1 to TP 2
TP1.2_Growth.plot<-TP1.2_Growth.plot+ ggtitle("TP1 to TP2")+
  theme(plot.title = element_text(colour="black", 
                                  size=panel.lab.sz, hjust=0.5), 
        legend.position=c(0.25, 0.85), legend.direction="horizontal")+
  labs(colour=NULL)+
  guides(color=guide_legend(nrow=2, byrow=FALSE)) 


##TP 3 to TP 4
TP3.4_Growth.plot<-TP3.4_Growth.plot+ ggtitle("TP3 to TP4")+
  theme(plot.title = element_text(colour="black", 
                                  size=panel.lab.sz, hjust=0.5), 
        legend.position="none")

```


#### Figure S5
```{r}
##Create Panel
Growth_fig<-plot_grid(TP1.2_Growth.plot, TP3.4_Growth.plot, 
                    rel_widths=1, rel_heights = 1,
                    nrow=1, ncol=2, byrow=T, labels = c("A", "B"), 
                    label_size=panel.lab.sz)

##Save Figure
ggsave(filename="Figures/03_Performance/FigS5_Growth.png", plot=Growth_fig, dpi=300, width=10, height=4.5, units="in")

```


### Figure S6 Initial Heat Assay

#### Adjust Figures for Panel
```{r}
##FvFm
IN_PAM.plot<-IN_PAM.plot + labs(x="", y="Fv/Fm") + theme(legend.position="none") + ylim(0.5, 0.655)

IN_TolPAM.plot<-IN_TolPAM.plot + labs(x="", y="Retention") + 
  theme(legend.position="none") + ylim(0.75, 1)

##Symbionts
IN_Sym.plot<-IN_Sym.plot + labs(x="", y="Symbionts") + theme(legend.position="none") + ylim(0,1)

IN_TolSym.plot<-IN_TolSym.plot + labs(x="", y="Retention") + 
  theme(legend.position="none") + ylim(.25, .75)

##Chlorophyll
IN_Chl.plot<-IN_Chl.plot + labs(x="Site and Treatment", y="Chlorophyll") + 
  theme(legend.position="none") + ylim(0,1.25)

IN_TolChl.plot<-IN_TolChl.plot + labs(x="Site", y="Retention") + 
  theme(legend.position="none") + ylim(0, 0.25)

```


#### Figure S6
```{r}
##Create Panel
IN_HeatAssay_fig<-plot_grid(IN_PAM.plot, IN_TolPAM.plot, 
                      IN_Sym.plot, IN_TolSym.plot,
                      IN_Chl.plot, IN_TolChl.plot,
                    rel_widths=c(1, 0.75), 
                    rel_heights = c(1, 1, 1),
                    nrow=3, ncol=2, byrow=T, 
                    labels = c("A", "B", "C", "D", "E", "F"), 
                    label_size=panel.lab.sz)

##Save Figure
ggsave(filename="Figures/03_Performance/FigS6_InitialHeatAssay.png", plot=IN_HeatAssay_fig, dpi=300, width=8, height=11, units="in")

```


### Figure S7 Thermal Tolerance

#### Adjust Figures for Panel
```{r}
##FvFm Retention
TP1_TolPAM.plot<-TP1_TolPAM.plot + ggtitle("TP1")+ 
  labs(x=NULL, y="Photo. Effic. Retention", colour=NULL)+
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position=c(0.5, 0.15), legend.direction="horizontal")+
  guides(color=guide_legend(nrow=2, byrow=FALSE))

TP2_TolPAM.plot<-TP2_TolPAM.plot+ ggtitle("TP2")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")

TP3_TolPAM.plot<-TP3_TolPAM.plot+ ggtitle("TP3")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")

TP4_TolPAM.plot<-TP4_TolPAM.plot+ ggtitle("TP4")+ labs(x=NULL, y=NULL) +
  theme(plot.title = element_text(colour="black", size=panel.lab.sz, face="bold", hjust=0.5), 
        legend.position="none")


##Symbiont Retention
TP1_TolSym.plot<- TP1_TolSym.plot + labs(x=NULL) + theme(legend.position="none")

TP2_TolSym.plot<- TP2_TolSym.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")

TP3_TolSym.plot<- TP3_TolSym.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")

TP4_TolSym.plot<- TP4_TolSym.plot + labs(x=NULL, y=NULL) + theme(legend.position="none")


##Chlorophyll Retention
TP1_TolChl.plot<-TP1_TolChl.plot + theme(legend.position="none")

TP2_TolChl.plot<-TP2_TolChl.plot + labs(y=NULL) + theme(legend.position="none")

TP3_TolChl.plot<-TP3_TolChl.plot + labs(y=NULL) + theme(legend.position="none")

TP4_TolChl.plot<-TP4_TolChl.plot + labs(y=NULL) + theme(legend.position="none")

```


#### Figure S7
```{r}
##Create Panel
Tolerance_fig<-plot_grid(TP1_TolPAM.plot, TP2_TolPAM.plot, 
                      TP3_TolPAM.plot, TP4_TolPAM.plot, 
                      TP1_TolSym.plot, TP2_TolSym.plot, 
                      TP3_TolSym.plot, TP4_TolSym.plot, 
                      TP1_TolChl.plot, TP2_TolChl.plot, 
                      TP3_TolChl.plot, TP4_TolChl.plot, 
                    rel_widths=1, rel_heights = 1,
                    nrow=3, ncol=4, byrow=T, labels = NULL)

##Save Figure
ggsave(filename="Figures/03_Performance/FigS7_Tolerance.png", plot=Tolerance_fig, dpi=300, width=14, height=12, units="in")

```


# Tables

### Table S6 Performance LM
```{r}
##Save Results Table
TableS6_Perf.LM<-Perf.res
  
##Organize
names(TableS6_Perf.LM)
TableS6_Perf.LM<-TableS6_Perf.LM %>% dplyr::rename(SE = Std..Error, DF = df, t = t.value, Timepoint = TimeP)
TableS6_Perf.LM<-TableS6_Perf.LM[,c("Timepoint", "Response", "Predictor", "Estimate", "SE", "DF", "t", "p", "EtaSq")]

#Round to 3 digits
TableS6_Perf.LM$Estimate<-round(TableS6_Perf.LM$Estimate, 3)
TableS6_Perf.LM$SE<-round(TableS6_Perf.LM$SE, 3)
TableS6_Perf.LM$t<-round(TableS6_Perf.LM$t, 3)
TableS6_Perf.LM$p<-round(TableS6_Perf.LM$p, 3)
TableS6_Perf.LM$EtaSq<-round(TableS6_Perf.LM$EtaSq, 3)

#Integer
TableS6_Perf.LM$DF<-round(TableS6_Perf.LM$DF, 0)

##Write Out Table
write.csv(TableS6_Perf.LM, "Tables/TableS6_Performance_LM_Results.csv", row.names=FALSE)

```


### Table S7 Performance Pairwise
```{r}
##Save Results Table
TableS7_Perf.pair<-Perf.pair.res

##Organize
names(TableS7_Perf.pair)
TableS7_Perf.pair<-TableS7_Perf.pair %>% dplyr::rename(Timepoint = TimeP, Contrast =contrast, Estimate = estimate, DF = df, t = t.ratio, p = p.value, EffectSize = effect.size, EffectSize.SE = SE.es)
TableS7_Perf.pair<-TableS7_Perf.pair[,c("Timepoint", "Response", "Site", "Contrast", "Estimate", "SE", "DF", "t", "p", "EffectSize", "EffectSize.SE")]

#Round to 3 digits
TableS7_Perf.pair$Estimate<-round(TableS7_Perf.pair$Estimate, 3)
TableS7_Perf.pair$SE<-round(TableS7_Perf.pair$SE, 3)
TableS7_Perf.pair$t<-round(TableS7_Perf.pair$t, 3)
TableS7_Perf.pair$p<-round(TableS7_Perf.pair$p, 3)
TableS7_Perf.pair$EffectSize<-round(TableS7_Perf.pair$EffectSize, 3)
TableS7_Perf.pair$EffectSize.SE<-round(TableS7_Perf.pair$EffectSize.SE, 3)

#Integer
TableS7_Perf.pair$DF<-round(TableS7_Perf.pair$DF, 0)

##Write Out Table
write.csv(TableS7_Perf.pair, "Tables/TableS7_Performance_Pairwise_Results.csv", row.names=FALSE)

```


### Table S8 Initial Heat Assay LM
```{r}
##Combine Results Tables
TableS8_IN_Tolerance<-data.frame(rbind(PAM_IN.res, Sym_IN.res, Chl_IN.res))

##Organize
names(TableS8_IN_Tolerance)
TableS8_IN_Tolerance<-TableS8_IN_Tolerance %>% dplyr::rename(Estimate = estimate, DF = df, t = t.ratio, p = p.value, EffectSize = EtaSq)
TableS8_IN_Tolerance<-TableS8_IN_Tolerance[,c("Response", "Predictor", "Estimate", "SE", "DF", "t", "p", "EffectSize")]

#Round to 3 digits
TableS8_IN_Tolerance$Estimate<-round(TableS8_IN_Tolerance$Estimate, 3)
TableS8_IN_Tolerance$SE<-round(TableS8_IN_Tolerance$SE, 3)
TableS8_IN_Tolerance$t<-round(TableS8_IN_Tolerance$t, 3)
TableS8_IN_Tolerance$p<-round(TableS8_IN_Tolerance$p, 3)
TableS8_IN_Tolerance$EffectSize<-round(TableS8_IN_Tolerance$EffectSize, 3)

#Integer
TableS8_IN_Tolerance$DF<-round(TableS8_IN_Tolerance$DF, 0)

##Write Out Table
write.csv(TableS8_IN_Tolerance, "Tables/TableS8_Initial_Tolerance_LM_Results.csv", row.names=FALSE)

```
